% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.r
\name{app_server}
\alias{app_server}
\title{Main UI of IDBac}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}
}
\value{
IDBac server
}
\description{
Main UI of IDBac
}
