% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_appendToDendrogramLabels.R
\name{appendDendLabsUI}
\alias{appendDendLabsUI}
\title{Append metadata to the dendrogrm's labels}
\usage{
appendDendLabsUI(id)
}
\arguments{
\item{id}{namespace id}
}
\value{
shiny UImodule
}
\description{
Append metadata to the dendrogrm's labels
}
