% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{as.binary.matrix.hclust}
\alias{as.binary.matrix.hclust}
\title{as.binary.matrix.hclust}
\usage{
as.binary.matrix.hclust(x)
}
\arguments{
\item{x}{see: https://github.com/sgibb/bootstrap}
}
\value{
see: https://github.com/sgibb/bootstrap
}
\description{
as.binary.matrix.hclust
}
