% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleMirrorPlots.R
\name{assembleMirrorPlots}
\alias{assembleMirrorPlots}
\title{Assemble mirror plot Data}
\usage{
assembleMirrorPlots(sampleID1 = input$Spectra1,
  sampleID2 = input$Spectra2,
  peakPercentPresence = input$percentPresence,
  lowerMassCutoff = input$lowerMass, upperMassCutoff = input$upperMass,
  minSNR = input$SNR, pool = workingDB$pool())
}
\arguments{
\item{sampleID1}{sample ID to search in IDBac sqlite database (will be positive spectrum)}

\item{sampleID2}{sample ID to search in IDBac sqlite database (will be negative spectrum)}

\item{peakPercentPresence}{numeric between 0% and 100%, peakPercentPresence}

\item{lowerMassCutoff}{lowerMassCutoff}

\item{upperMassCutoff}{upperMassCutoff}

\item{minSNR}{numeric, peaks with a SNR below this number will be removed}

\item{pool}{IDBac database pool}
}
\value{
environment containing mirror plot data
}
\description{
Assemble mirror plot Data
}
