% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableSampleNames.R
\name{availableSampleNames}
\alias{availableSampleNames}
\title{Search an IDBac database to see which sample IDs have protein or small molecule data}
\usage{
availableSampleNames(checkedPool, whetherProtein, allSamples)
}
\arguments{
\item{whetherProtein}{}

\item{pool}{}
}
\description{
Search an IDBac database to see which sample IDs have protein or small molecule data
}
