% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mirrorPlots.R
\name{baserMirrorPlot}
\alias{baserMirrorPlot}
\title{Create mirror plot in base R}
\usage{
baserMirrorPlot(mirrorPlotEnv)
}
\arguments{
\item{mirrorPlotEnv}{data for plot}
}
\value{
NA
}
\description{
Create mirror plot in base R
}
