% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binnR.r
\name{binnR}
\alias{binnR}
\title{binnR}
\usage{
binnR(vectorList, ppm, refSeqStart, refSeqEnd)
}
\arguments{
\item{vectorList}{NA}

\item{ppm}{NA}

\item{refSeqStart}{NA}

\item{refSeqEnd}{NA}
}
\value{
NA
}
\description{
binnR
}
\examples{
NA
}
