% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootlabels.hclust}
\alias{bootlabels.hclust}
\title{Print bootstrap values.}
\usage{
bootlabels.hclust(x, bootstrapValues, horiz = FALSE, ...)
}
\arguments{
\item{x}{\code{hclust} object}

\item{bootstrapValues}{\code{numeric}, bootstrap values}

\item{horiz}{print values for a horizontal tree?}

\item{...}{see: https://github.com/sgibb/bootstrap}
}
\description{
This function prints bootstrap values to the corresponding node.
}
\seealso{
\code{\link{bootstrap}}
}
