% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\docType{package}
\name{bootstrap-package}
\alias{bootstrap-package}
\title{Simple Bootstrapping for Hierarchical Clustering}
\description{
This package provides bootstrapping for hierarchical clustering.
}
\details{
\tabular{ll}{
Package: \tab bootstrap \cr
Version: \tab 0.1\cr
Date: \tab 2013-06-27\cr
License: \tab GPL (>= 3)\cr
URL: \tab http://www.github.com/sgibb/bootstrap/ \cr
}
}
\references{
\url{http://www.github.com/sgibb/bootstrap/}
}
\author{
Sebastian Gibb <\email{mail@sebastiangibb.de}>
}
\keyword{package}
