% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap}
\usage{
bootstrap(x, fun, n = 1000L)
}
\arguments{
\item{x}{see: https://github.com/sgibb/bootstrap}

\item{fun}{see: https://github.com/sgibb/bootstrap}

\item{n}{see: https://github.com/sgibb/bootstrap}
}
\value{
see: https://github.com/sgibb/bootstrap
}
\description{
Bootstrap
}
