% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_shinyBS.R
\name{bsCollapse}
\alias{bsCollapse}
\title{Title}
\usage{
bsCollapse(..., id = NULL, multiple = FALSE, open = NULL)
}
\arguments{
\item{...}{.}

\item{id}{.}

\item{multiple}{.}

\item{open}{.}
}
\value{
.
}
\description{
Title
}
