% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_shinyBS.R
\name{bsCollapsePanel}
\alias{bsCollapsePanel}
\title{Title}
\usage{
bsCollapsePanel(title, ..., value = title, style = NULL)
}
\arguments{
\item{title}{.}

\item{...}{.}

\item{value}{..}

\item{style}{.}
}
\value{
.
}
\description{
Title
}
