% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_coloringDendLines.R
\name{changeDendLinesColor}
\alias{changeDendLinesColor}
\title{changeDendLinesColor}
\usage{
changeDendLinesColor(dendrogram, colorBy, colorBlindPalette,
  cutHeight = 0, chosenK = 1)
}
\arguments{
\item{dendrogram}{NA}

\item{colorBy}{NA}

\item{colorBlindPalette}{NA}

\item{cutHeight}{NA}

\item{chosenK}{NA}
}
\value{
NA
}
\description{
changeDendLinesColor
}
\examples{
NA
}
