% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_coloringDendLines.R
\name{changeDendLinesWidth}
\alias{changeDendLinesWidth}
\title{changeDendLinesWidth}
\usage{
changeDendLinesWidth(dendrogram, width)
}
\arguments{
\item{dendrogram}{NA}

\item{width}{NA}
}
\value{
NA
}
\description{
changeDendLinesWidth
}
\examples{
NA
}
