% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogramModifiers.R
\name{changeDendPartColor}
\alias{changeDendPartColor}
\title{changeDendPartColor}
\usage{
changeDendPartColor(dendrogram, colorBy, colorBlindPalette,
  cutHeight = 0, chosenK = 1, part)
}
\arguments{
\item{dendrogram}{a}

\item{colorBy}{a}

\item{colorBlindPalette}{a}

\item{cutHeight}{a}

\item{chosenK}{a}

\item{part}{a}
}
\value{
a
}
\description{
changeDendPartColor
}
