% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogramModifiers.R
\name{changeDendPartSize}
\alias{changeDendPartSize}
\title{changeDendPartSize}
\usage{
changeDendPartSize(dendrogram, dendPartSize, part)
}
\arguments{
\item{dendrogram}{a}

\item{dendPartSize}{a}

\item{part}{a}
}
\value{
a
}
\description{
changeDendPartSize
}
