% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chartoRawtoCompressed}
\alias{chartoRawtoCompressed}
\title{Take character, turn to raw, then compress (note: base::charToRaw is not vectorized)}
\usage{
chartoRawtoCompressed(input, compression)
}
\arguments{
\item{input}{character}

\item{compression}{}
}
\description{
Take character, turn to raw, then compress (note: base::charToRaw is not vectorized)
}
