% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_chooseDirectory.r
\name{choose_dir}
\alias{choose_dir}
\title{Choose Directory, choose.dir vs tcltk}
\usage{
choose_dir(caption = "Select data directory")
}
\arguments{
\item{caption}{NA}
}
\value{
NA
}
\description{
Choose Directory, choose.dir vs tcltk
}
