% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooserInput.r
\name{chooserInput}
\alias{chooserInput}
\title{chooserInput}
\usage{
chooserInput(inputId, leftLabel, rightLabel, leftChoices, rightChoices,
  size = 5, multiple = FALSE)
}
\arguments{
\item{inputId}{NA}

\item{leftLabel}{NA}

\item{rightLabel}{NA}

\item{leftChoices}{NA}

\item{rightChoices}{NA}

\item{size}{NA}

\item{multiple}{NA}
}
\value{
NA
}
\description{
chooserInput
}
