% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanWSpace}
\alias{cleanWSpace}
\title{People have trouble with spaces}
\usage{
cleanWSpace(input)
}
\arguments{
\item{input}{character vector}
}
\value{
character vector
}
\description{
People have trouble with spaces
}
