% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseReplicates.R
\name{collapseReplicates}
\alias{collapseReplicates}
\title{Collapse a sample's MALDIquant peak objects into a single peak object}
\usage{
collapseReplicates(
  pool,
  sampleIDs,
  peakPercentPresence,
  lowerMassCutoff,
  upperMassCutoff,
  minSNR,
  tolerance = 0.002,
  protein
)
}
\arguments{
\item{pool}{sqlite pool}

\item{sampleIDs}{sample IDs of samples to process}

\item{peakPercentPresence}{peaks in replciates that occurr less frequently than this will be removed}

\item{lowerMassCutoff}{masses below this will be removed from analyses}

\item{upperMassCutoff}{masses above this will be removed from analyses}

\item{minSNR}{minimum SNR a a peak must have to be retained}

\item{tolerance}{binning tolerance ppm / 10e6}

\item{protein}{whether to search SQL for protein or small mol spectra}
}
\value{
a single trimmed and binned MALDIquant peak object
}
\description{
Collapse a sample's MALDIquant peak objects into a single peak object
}
