% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlManipulators.r
\name{collapseSmallMolReplicates}
\alias{collapseSmallMolReplicates}
\title{collapseSmallMolReplicates}
\usage{
collapseSmallMolReplicates(db, fileshas, smallMolPercentPresence,
  lowerMassCutoff, upperMassCutoff)
}
\arguments{
\item{db}{NA}

\item{fileshas}{NA}

\item{smallMolPercentPresence}{NA}

\item{lowerMassCutoff}{NA}

\item{upperMassCutoff}{NA}
}
\value{
NA
}
\description{
collapseSmallMolReplicates
}
\examples{
NA
}
