% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_modDendrogramLabels.R
\name{colordendLabels}
\alias{colordendLabels}
\title{colordendLabels}
\usage{
colordendLabels(input, output, session, dendrogram)
}
\arguments{
\item{input}{a}

\item{output}{a}

\item{session}{a}

\item{dendrogram}{a}
}
\value{
a
}
\description{
colordendLabels
}
