% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_modifyDendrogramLabels.R
\name{colordendLabelsUI}
\alias{colordendLabelsUI}
\title{Modify (color/size) dendrogram Labels}
\usage{
colordendLabelsUI(id)
}
\arguments{
\item{id}{namespace id}
}
\value{
ui for coloring dendrogram labels
}
\description{
Modify (color/size) dendrogram Labels
}
