% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_modDendrogramLines.R
\name{colordendLines}
\alias{colordendLines}
\title{colordendLines}
\usage{
colordendLines(input, output, session, dendrogram, plotHeight = 500)
}
\arguments{
\item{input}{a}

\item{output}{a}

\item{session}{a}

\item{dendrogram}{a}

\item{plotHeight}{a}
}
\value{
a
}
\description{
colordendLines
}
