% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_coloringDendLines.R
\name{colordendLinesServer}
\alias{colordendLinesServer}
\title{colordendLinesServer}
\usage{
colordendLinesServer(input, output, session, dendrogram,
  plotHeight = 500)
}
\arguments{
\item{input}{NA}

\item{output}{NA}

\item{session}{NA}

\item{dendrogram}{NA}

\item{plotHeight}{NA}
}
\value{
NA
}
\description{
colordendLinesServer
}
\examples{
NA
}
