% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_coloringDendLines.R
\name{colordendLinesServercolordendLinesUI}
\alias{colordendLinesServercolordendLinesUI}
\title{colordendLinesServercolordendLinesUI}
\usage{
colordendLinesServercolordendLinesUI(id)
}
\arguments{
\item{id}{NA}
}
\value{
NA
}
\description{
colordendLinesServercolordendLinesUI
}
\examples{
NA
}
