% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_ModifyDendrogramLines.R
\name{colordendLinesUI}
\alias{colordendLinesUI}
\title{colordendLinesUI}
\usage{
colordendLinesUI(id)
}
\arguments{
\item{id}{namespace id}
}
\value{
shiny UImodule
}
\description{
colordendLinesUI
}
