% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colored_Dots.r
\name{colored_dots}
\alias{colored_dots}
\title{colored_dots
I submitted this and it was pulled into dendextend but is left here for less dependency/more control}
\usage{
colored_dots(colors, dend, rowLabels = NULL, cex.rowLabels = 0.9,
  add = TRUE, y_scale, y_shift, text_shift = 1,
  sort_by_labels_order = TRUE, horiz = FALSE, alf = 1,
  circ = FALSE, ...)
}
\arguments{
\item{colors}{NA}

\item{dend}{NA}

\item{rowLabels}{NA}

\item{cex.rowLabels}{NA}

\item{add}{NA}

\item{y_scale}{NA}

\item{y_shift}{NA}

\item{text_shift}{NA}

\item{sort_by_labels_order}{NA}

\item{horiz}{NA}

\item{alf}{NA}

\item{circ}{NA}

\item{...}{NA}
}
\value{
NA
}
\description{
colored_dots
I submitted this and it was pulled into dendextend but is left here for less dependency/more control
}
