% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dendrogram.R
\name{colorPicker}
\alias{colorPicker}
\title{colorPicker for dend and dots}
\usage{
colorPicker(levs, ns)
}
\arguments{
\item{levs}{levels (reactiveValue)}

\item{ns}{shiny namespace}
}
\value{
list of html for each level with colors chosen
}
\description{
colorPicker for dend and dots
}
