% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compress}
\alias{compress}
\title{compress
  Settings for compressing in IDBac}
\usage{
compress(input)
}
\arguments{
\item{input}{character or raw vector to compressed (char will be converted to raw)}
}
\value{
Raw vector
}
\description{
compress
  Settings for compressing in IDBac
}
