% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_conversion.R
\name{controlBrukerDisplay}
\alias{controlBrukerDisplay}
\title{Controls display of Bruker conversion page}
\usage{
controlBrukerDisplay(session, pwizAvailable, ostest = NULL)
}
\arguments{
\item{session}{shiny session}

\item{pwizAvailable}{whether msconvert was found}

\item{ostest}{for testing function}
}
\value{
html
}
\description{
Controls display of Bruker conversion page
}
