% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_conversion.R
\name{convertDataTab_Server}
\alias{convertDataTab_Server}
\title{convertDataTab_Server}
\usage{
convertDataTab_Server(input, output, session, tempMZDir, sqlDirectory)
}
\arguments{
\item{input}{.}

\item{output}{.}

\item{session}{.}

\item{tempMZDir}{.}

\item{sqlDirectory}{.}
}
\value{
.
}
\description{
convertDataTab_Server
}
