% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_convertDelim.R
\name{convertDelim_Server}
\alias{convertDelim_Server}
\title{convertDelim_Server}
\usage{
convertDelim_Server(
  input,
  output,
  session,
  tempMZDir,
  sqlDirectory,
  availableExperiments,
  ...
)
}
\arguments{
\item{input}{module}

\item{output}{module}

\item{session}{module}

\item{tempMZDir}{tempMZDir}

\item{sqlDirectory}{sqlDirectory}

\item{availableExperiments}{availableExperiments}

\item{...}{advanced arguments for MALDIquant, see \code{\link[IDBacApp:processSmallMolSpectra]{IDBacApp::processSmallMolSpectra()}} and/or \code{\link[IDBacApp:processProteinSpectra]{IDBacApp::processProteinSpectra()}}}
}
\value{
.
}
\description{
convertDelim_Server
}
