% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_convertMicrotyper.R
\name{convertMicrotyper_Server}
\alias{convertMicrotyper_Server}
\title{convertDelim_Server}
\usage{
convertMicrotyper_Server(input, output, session, tempMZDir, sqlDirectory)
}
\arguments{
\item{input}{module}

\item{output}{module}

\item{session}{module}

\item{tempMZDir}{tempMZDir}

\item{sqlDirectory}{sqlDirectory}
}
\value{
.
}
\description{
convertDelim_Server
}
