% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_convertMZ.R
\name{convertMZ_Server}
\alias{convertMZ_Server}
\title{convertMZ_Server}
\usage{
convertMZ_Server(input, output, session, sqlDirectory, availableExperiments)
}
\arguments{
\item{input}{module}

\item{output}{module}

\item{session}{module}

\item{sqlDirectory}{sqlDirectory}

\item{availableExperiments}{availableExperiments}
}
\value{
.
}
\description{
convertMZ_Server
}
