% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_convertBruker.R
\name{convertOneBruker_Server}
\alias{convertOneBruker_Server}
\title{#' oneMaldiPlateHelpUI
#'
#' @param id NA
#'
#' @return NA
#' @export
#'
#' @examples NA
oneMaldiPlateHelpUI <- function(id){
  ns <- shiny::NS(id)
  tagList(
    h3("Instructions", align = "center"),
    
    br(),
    p(strong("1: Working Directory"), " This directs where on your computer you would like to create an IDBac working directory."),
    p("In the folder you select, IDBac will create sub-folders within a main directory named \"IDBac\":"),
    img(src = "WorkingDirectory.png",
        style = "width:60%;height:60%"),
    br(),
    p(strong("2: Raw Data"), "Your RAW data is a single folder that contains: one subfolder containing protein
      data and one subfolder containing small-molecule data"),
    img(src = "Single-MALDI-Plate.png",
        style = "width:60%;height:60%"),
    br(),
    p("*Note: Sometimes the browser window won't pop up, but will still appear in the application bar. See below:"),
    img(src = "window.png",
        width = "100%")
    )
  
}}
\usage{
convertOneBruker_Server(input, output, session, tempMZDir, sqlDirectory)
}
\description{
#' oneMaldiPlateHelpUI
#'
#' @param id NA
#'
#' @return NA
#' @export
#'
#' @examples NA
oneMaldiPlateHelpUI <- function(id){
  ns <- shiny::NS(id)
  tagList(
    h3("Instructions", align = "center"),
    
    br(),
    p(strong("1: Working Directory"), " This directs where on your computer you would like to create an IDBac working directory."),
    p("In the folder you select, IDBac will create sub-folders within a main directory named \"IDBac\":"),
    img(src = "WorkingDirectory.png",
        style = "width:60%;height:60%"),
    br(),
    p(strong("2: Raw Data"), "Your RAW data is a single folder that contains: one subfolder containing protein
      data and one subfolder containing small-molecule data"),
    img(src = "Single-MALDI-Plate.png",
        style = "width:60%;height:60%"),
    br(),
    p("*Note: Sometimes the browser window won't pop up, but will still appear in the application bar. See below:"),
    img(src = "window.png",
        width = "100%")
    )
  
}
}
