% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToMZML.r
\name{convertToMzml}
\alias{convertToMzml}
\title{Convert to mzML}
\usage{
convertToMzml(mzmlRawFileDirectory, mzmlRawFilesLocation,
  msconvertLocation, outDir, spectraConversion)
}
\arguments{
\item{mzmlRawFileDirectory}{NA}

\item{mzmlRawFilesLocation}{NA}

\item{msconvertLocation}{NA}

\item{outDir}{NA}

\item{spectraConversion}{NA}
}
\value{
NA
}
\description{
Convert to mzML
}
\examples{
NA
}
