% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyToNewDatabase.r
\name{copyDB_dbDetach}
\alias{copyDB_dbDetach}
\title{Detach new database to existing database}
\usage{
copyDB_dbDetach(newdbPath, existingDBconn)
}
\arguments{
\item{newdbPath}{newdbPath}

\item{existingDBconn}{existingDBconn}
}
\value{
NA
}
\description{
Detach new database to existing database
}
