% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyToNewDatabase.r
\name{copyDB_setupIndividualSpectra}
\alias{copyDB_setupIndividualSpectra}
\title{Setup IndividualSpectra DB table}
\usage{
copyDB_setupIndividualSpectra(newDBconn, existingDBconn, arch)
}
\arguments{
\item{newDBconn}{newDBconn}

\item{existingDBconn}{existingDBconn}

\item{arch}{DB architecture}
}
\value{
NA
}
\description{
Setup IndividualSpectra DB table
}
