% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyToNewDatabase.r
\name{copyDB_setupMeta}
\alias{copyDB_setupMeta}
\title{Setup metadata DB table}
\usage{
copyDB_setupMeta(newDBconn, existingDBconn, arch)
}
\arguments{
\item{newDBconn}{newDBconn}

\item{existingDBconn}{existingDBconn}

\item{arch}{DB architecture}
}
\value{
NA
}
\description{
Setup metadata DB table
}
