% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyToNewDatabase.r
\name{copyDB_setupXML}
\alias{copyDB_setupXML}
\title{Setup XML DB table}
\usage{
copyDB_setupXML(newDBconn, existingDBconn, arch)
}
\arguments{
\item{newDBconn}{newDBconn}

\item{existingDBconn}{existingDBconn}

\item{arch}{DB architecture}
}
\value{
NA
}
\description{
Setup XML DB table
}
