% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_popups.R
\name{copyingDbPopup}
\alias{copyingDbPopup}
\title{While Database is copying, block user-interaction}
\usage{
copyingDbPopup()
}
\value{
modal
}
\description{
While Database is copying, block user-interaction
}
