% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_peakFuzzifier.R
\name{createFuzzyVectorUnit}
\alias{createFuzzyVectorUnit}
\title{Experimental binning AKA- not for use}
\usage{
createFuzzyVectorUnit(massStart, massEnd, chunksize, massList)
}
\arguments{
\item{massStart}{beginning of mass range (as of now, must be smaller than smallest mass)}

\item{massEnd}{end of mass range (as of now, must be smaller than smallest mass)}

\item{chunksize}{size of bins}

\item{massList}{list of mass vectors (eg list(1:10, 1:10))}
}
\value{
matrix where rows are samples and columns are variables (m/z preojections)
}
\description{
Experimental binning AKA- not for use
}
\examples{
\dontrun{
masses <- list(Sample_A = c(5000,6000,7000), 
Sample_B = c(5000,6010,7005), Sample_C = c(5000,6010,7005))  
intensities <- list(Sample_A = rep(1, 3), 
Sample_B = rep(1, 3),
 Sample_C = rep(1, 3))
zx <- binnR(massStart = 3000,
            massEnd = 15000,
            ppm = 500,
            massList = masses,
            intensityList = intensities)
            }
}
