% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_dbCreators.R
\name{createMetaSQL}
\alias{createMetaSQL}
\title{createMetaSQL}
\usage{
createMetaSQL(sampleID, pool)
}
\arguments{
\item{sampleID}{NA}

\item{pool}{sqlite pool}
}
\value{
writes to sqlite database
}
\description{
createMetaSQL
}
