% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPool.R
\name{createNewSQLITEdb}
\alias{createNewSQLITEdb}
\title{Create New, Empty SQLite Database}
\usage{
createNewSQLITEdb(newExperimentName, sqlDirectory)
}
\arguments{
\item{newExperimentName}{name for the sqlite DB}

\item{sqlDirectory}{directory to which the sqlite DB will be written into}
}
\value{
pool connection
}
\description{
Create New, Empty SQLite Database
}
