% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbCreators.R
\name{createSpectraSQL}
\alias{createSpectraSQL}
\title{createSpectraSQL}
\usage{
createSpectraSQL(mzML_con, scanNumber, userDBCon, sampleID, XMLinfo,
  rawDataFilePath, smallRangeEnd = 6000)
}
\arguments{
\item{mzML_con}{NA}

\item{scanNumber}{NA}

\item{userDBCon}{NA}

\item{sampleID}{NA}

\item{XMLinfo}{NA}

\item{rawDataFilePath}{NA}

\item{smallRangeEnd}{end of mass region for small mol, if m/z above this- will be classified as "protein" spectrum}
}
\value{
NA
}
\description{
createSpectraSQL
}
