% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createSpectrumSha}
\alias{createSpectrumSha}
\title{createSpectrumSha
   Given a mzR}
\usage{
createSpectrumSha(peaklist)
}
\arguments{
\item{peaklist}{matrix}
}
\value{
sha
}
\description{
createSpectrumSha
   Given a mzR
}
