% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_dbCreators.R
\name{createXMLSQL}
\alias{createXMLSQL}
\title{createXMLSQL}
\usage{
createXMLSQL(rawDataFilePath, pool, mzML_con)
}
\arguments{
\item{rawDataFilePath}{xml path}

\item{pool}{sqlite pool}

\item{mzML_con}{mzR connection}
}
\value{
writes to sqlite database
}
\description{
createXMLSQL
}
