% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_current_database_version.R
\name{current_db_version}
\alias{current_db_version}
\title{Single point of reference for latest db version}
\usage{
current_db_version()
}
\value{
character semantic version
}
\description{
Single point of reference for latest db version
}
