% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_databaseSelector.R
\name{databaseSelector_server}
\alias{databaseSelector_server}
\title{databaseSelector_server}
\usage{
databaseSelector_server(input, output, session,
  h3Label = "First, select an experiment:", availableExperiments,
  sqlDirectory)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{h3Label}{h3Label}

\item{availableExperiments}{availableExperiments}
}
\value{
filepath of selected database
}
\description{
databaseSelector_server
}
