% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_databaseSelector.R
\name{databaseSelector_UI}
\alias{databaseSelector_UI}
\title{databaseSelector_UI}
\usage{
databaseSelector_UI(id)
}
\arguments{
\item{id}{namespace}
}
\value{
modal ui
}
\description{
databaseSelector_UI
}
