% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_databaseTab.R
\name{databaseTabServer}
\alias{databaseTabServer}
\title{databaseTabServer}
\usage{
databaseTabServer(input, output, session, sqlDirectory, availableExperiments)
}
\arguments{
\item{input}{module}

\item{output}{module}

\item{session}{module}

\item{sqlDirectory}{sqlDirectory}

\item{availableExperiments}{availableExperiments}
}
\value{
.
}
\description{
databaseTabServer
}
