% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_databaseTab.R
\name{databaseTabUI}
\alias{databaseTabUI}
\title{Layout for the databaseTabUI}
\usage{
databaseTabUI(id)
}
\arguments{
\item{id}{namespace}
}
\value{
ui for database UI
}
\description{
Layout for the databaseTabUI
}
