% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popups.R
\name{dbExists_server}
\alias{dbExists_server}
\title{Modular modal Warn user that database alredy exists}
\usage{
dbExists_server(input, output, session, continue)
}
\arguments{
\item{input}{shiny}

\item{output}{shiny}

\item{session}{shiny}
}
\value{
modular modal
}
\description{
Modular modal Warn user that database alredy exists
}
