% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popups.R
\name{dbExists_UI}
\alias{dbExists_UI}
\title{Modular modal Warn user that database alredy exists}
\usage{
dbExists_UI(id, dbName)
}
\arguments{
\item{id}{namespace}

\item{dbName}{name of database to be created}
}
\value{
modular modal
}
\description{
Modular modal Warn user that database alredy exists
}
