% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{decompress}
\alias{decompress}
\title{decompress
Settings for decompressing in IDBac, serialized called first}
\usage{
decompress(input)
}
\arguments{
\item{input}{compressed raw vector}
}
\value{
raw vector
}
\description{
decompress
Settings for decompressing in IDBac, serialized called first
}
