% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_parseDelimitedMS.R
\name{delim_to_mzml}
\alias{delim_to_mzml}
\title{Parse Delimited MS files with fread and MALDIquant}
\usage{
delim_to_mzml(
  proteinPaths,
  proteinNames,
  smallMolPaths,
  smallMolNames,
  exportDirectory,
  centroid
)
}
\arguments{
\item{proteinPaths}{proteinPaths}

\item{proteinNames}{proteinNames}

\item{smallMolPaths}{smallMolPaths}

\item{smallMolNames}{smallMolNames names of the}

\item{exportDirectory}{Where will mzml files be written?}

\item{centroid}{Do the files contain profile or centroid data (currently implemented as just
one TRUE/FALSE, but could mappply)}
}
\value{
sample-named vector of filpaths of random-named mzml files
}
\description{
Parse Delimited MS files with fread and MALDIquant
}
