% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendAndDots.R
\name{dendAndDots}
\alias{dendAndDots}
\title{dendAndDots}
\usage{
dendAndDots(input, output, session, dendrogram)
}
\arguments{
\item{input}{default shiny}

\item{output}{default shiny}

\item{session}{default shiny}

\item{dendrogram}{dendrogram input (eg stats::as.dendrogram)}
}
\value{
NA
}
\description{
dendAndDots
}
