% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dendrogram.R
\name{dendDotsServer}
\alias{dendDotsServer}
\title{Title}
\usage{
dendDotsServer(input, output, session, dendrogram, pool, plotWidth,
  plotHeight, boots, dendOrPhylo = "Dendrogram")
}
\arguments{
\item{input}{NA}

\item{output}{NA}

\item{session}{NA}

\item{dendrogram}{NA}

\item{pool}{NA}

\item{plotWidth}{NA}

\item{plotHeight}{NA}

\item{boots}{bootstraps}

\item{dendOrPhylo}{should the dendrogram labels hang?}
}
\value{
NA
}
\description{
Title
}
