% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dend_and_dots.R
\name{dendDotsUI}
\alias{dendDotsUI}
\title{Modify (color/size) dendrogram lines}
\usage{
dendDotsUI(id)
}
\arguments{
\item{id}{namespace id}
}
\value{
shiny UImodule
}
\description{
Modify (color/size) dendrogram lines
}
