% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dendrogram.R
\name{dendrogramActionsUI}
\alias{dendrogramActionsUI}
\title{dendrogramActionsUI}
\usage{
dendrogramActionsUI(id)
}
\arguments{
\item{id}{namespace id}
}
\value{
shiny UImodule
}
\description{
dendrogramActionsUI
}
