% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dendrogramCreator.R
\name{dendrogramCreator}
\alias{dendrogramCreator}
\title{dendrogramCreator}
\usage{
dendrogramCreator(input, output, session, data)
}
\arguments{
\item{input}{a}

\item{output}{a}

\item{session}{a}

\item{data}{a}
}
\value{
a
}
\description{
dendrogramCreator
}
