% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dendrogramCreator.R
\name{dendrogramCreatorUI}
\alias{dendrogramCreatorUI}
\title{dendrogramCreatorUI}
\usage{
dendrogramCreatorUI(id)
}
\arguments{
\item{id}{namespace id}
}
\value{
ui to choose the algorithm for creating the dendrogram
}
\description{
dendrogramCreatorUI
}
