% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deserial}
\alias{deserial}
\title{deserial
   Settings for serializing in IDBac (convert to json)}
\usage{
deserial(input)
}
\arguments{
\item{input}{matrix or vector}
}
\value{
JSON
}
\description{
deserial
   Settings for serializing in IDBac (convert to json)
}
