% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dendrogram.R
\name{displayMissingProteinUI}
\alias{displayMissingProteinUI}
\title{Display samples removed from analysis}
\usage{
displayMissingProteinUI(id, sampleIds)
}
\arguments{
\item{id}{namespace}

\item{sampleIds}{sampleIds}
}
\value{
shiny ui
}
\description{
Display samples removed from analysis
}
