% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceMatrix.r
\name{distMatrix}
\alias{distMatrix}
\title{Create Distance Matrix}
\usage{
distMatrix(data, method, booled)
}
\arguments{
\item{data}{matrix or dataframe, where columns are variables and rows are samples}

\item{method}{distance metric to use}

\item{booled}{whether all values >0 should be set to one}
}
\value{
distance matrix
}
\description{
Create Distance Matrix
}
