% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{.calculateMatches}
\alias{.calculateMatches}
\title{.calculateMatches}
\usage{
.calculateMatches(origin, current, nc = ncol(origin))
}
\arguments{
\item{origin}{see: https://github.com/sgibb/bootstrap}

\item{current}{see: https://github.com/sgibb/bootstrap}

\item{nc}{see: https://github.com/sgibb/bootstrap}
}
\value{
see: https://github.com/sgibb/bootstrap
}
\description{
.calculateMatches
}
