% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.checkPool}
\alias{.checkPool}
\title{Check if a pool object}
\usage{
.checkPool(pool)
}
\arguments{
\item{pool}{variable to check}
}
\value{
NA, stops function if not a pool object
}
\description{
Check if a pool object
}
