% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{.clust}
\alias{.clust}
\title{.clust}
\usage{
.clust(x, fun)
}
\arguments{
\item{x}{see: https://github.com/sgibb/bootstrap}

\item{fun}{see: https://github.com/sgibb/bootstrap}
}
\value{
see: https://github.com/sgibb/bootstrap
}
\description{
.clust
}
