% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.db_path_from_pool}
\alias{.db_path_from_pool}
\title{Find database path from pool object}
\usage{
.db_path_from_pool(pool)
}
\arguments{
\item{pool}{{pool} object}
}
\value{
Path to pool's database
}
\description{
Find database path from pool object
}
