% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_updateIdbac.R
\name{.download_idbac_exe}
\alias{.download_idbac_exe}
\title{Download IDBac exe}
\usage{
.download_idbac_exe(latest_version_httr)
}
\arguments{
\item{latest_version_httr}{httr content}
}
\value{
none
}
\description{
Download IDBac exe
}
