% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_getSubTrees.R
\name{.get_one_subtree}
\alias{.get_one_subtree}
\title{Recursively look for a tree within a tree}
\usage{
.get_one_subtree(subdend, clusters)
}
\arguments{
\item{subdend}{dendrogram}

\item{clusters}{leaf indices of cluster}
}
\value{
dendrogram
}
\description{
Recursively look for a tree within a tree
}
