% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_updateIdbac.R
\name{.getLatestStableVersionPage}
\alias{.getLatestStableVersionPage}
\title{GET api.github.com page for latest release}
\usage{
.getLatestStableVersionPage(
  base_url = "https://api.github.com/repos/chasemc/IDBacApp/releases/latest"
)
}
\arguments{
\item{base_url}{url to GET}
}
\value{
httr content
}
\description{
GET api.github.com page for latest release
}
