% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_getSmallPeaksFromBrush.R
\name{.getSmallPeaksFromBrush}
\alias{.getSmallPeaksFromBrush}
\title{Retrieve small molecule and matrix peak lists and make consensus peak lists}
\usage{
.getSmallPeaksFromBrush(
  pool,
  sampleIDs = NULL,
  dendrogram,
  brushInputs,
  matrixIDs = NULL,
  peakPercentPresence,
  lowerMassCutoff,
  upperMassCutoff,
  minSNR
)
}
\arguments{
\item{pool}{pool}

\item{sampleIDs}{character vector of IDs (currently not used)}

\item{dendrogram}{protein dendrogram that was brushed}

\item{brushInputs}{brusked dendrogram outputs}

\item{matrixIDs}{sampleID that will be used as a matrix}

\item{peakPercentPresence}{peaks whose frequency of occurence in replicates below this are dropped}

\item{lowerMassCutoff}{lower mass to retain}

\item{upperMassCutoff}{upper mass to retain}

\item{minSNR}{minimum SNR a peak must have to be retained}
}
\value{
a list containg two lists of MALDIquant peak objects "samplePeaks" and "matrixPeaks"
}
\description{
Retrieve small molecule and matrix peak lists and make consensus peak lists
}
