% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readBrukerAcqusFile.r
\name{.grepAcquDoubleValue}
\alias{.grepAcquDoubleValue}
\title{.grepAcquDoubleValue}
\usage{
.grepAcquDoubleValue(patternStr, srcStr)
}
\arguments{
\item{patternStr}{NA}

\item{srcStr}{NA}
}
\value{
NA
}
\description{
.grepAcquDoubleValue
}
