% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBrukerAcqusFile.r
\name{.grepAcquValue}
\alias{.grepAcquValue}
\title{.grepAcquValue}
\usage{
.grepAcquValue(patternStr, srcStr)
}
\arguments{
\item{patternStr}{NA}

\item{srcStr}{NA}
}
\value{
NA
}
\description{
.grepAcquValue
}
\examples{
NA
}
