% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_idbacRetrievePeaksFromPool.R
\name{.parse_json_peaks}
\alias{.parse_json_peaks}
\title{Create maldiquant peaks from IDBac SQL peak JSON}
\usage{
.parse_json_peaks(input, minSNR)
}
\arguments{
\item{input}{IDBac SQL strain_id ancd peak json}

\item{minSNR}{minimum SNR a a peak must have to be retained}
}
\value{
list of maldiquant peak objects
}
\description{
Create maldiquant peaks from IDBac SQL peak JSON
}
