% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_bootstrap.R
\name{.resample}
\alias{.resample}
\title{resample}
\usage{
.resample(x, size = ncol(x))
}
\arguments{
\item{x}{see: https://github.com/sgibb/bootstrap}

\item{size}{see: https://github.com/sgibb/bootstrap}
}
\value{
see: https://github.com/sgibb/bootstrap
}
\description{
resample
}
