% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_peakFuzzifier.R
\name{.scale_ppm}
\alias{.scale_ppm}
\title{scaler so smallest mass differnce is 1}
\usage{
.scale_ppm(mass, ppm)
}
\arguments{
\item{mass}{numeric}

\item{ppm}{numeric}
}
\value{
numeric
}
\description{
scaler so smallest mass differnce is 1
}
