% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_bootstrap.R
\name{.text.coord.hclust}
\alias{.text.coord.hclust}
\title{.text.coord.hclust}
\usage{
.text.coord.hclust(x)
}
\arguments{
\item{x}{see: https://github.com/sgibb/bootstrap}
}
\value{
see: https://github.com/sgibb/bootstrap
}
\description{
.text.coord.hclust
}
