% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_idbac_double_from_raw_compressed.R
\name{double_from_raw_compressed}
\alias{double_from_raw_compressed}
\title{Convert raw compresed data from sqlite into R vector}
\usage{
double_from_raw_compressed(x)
}
\arguments{
\item{x}{raw compressed}
}
\value{

}
\description{
Convert raw compresed data from sqlite into R vector
}
