% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dendrogram.R
\name{downloadHier}
\alias{downloadHier}
\title{Download newick hierarchical dendrogram}
\usage{
downloadHier(id)
}
\arguments{
\item{id}{namespace}
}
\value{
NA
}
\description{
Download newick hierarchical dendrogram
}
