% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dendrogram.R
\name{downloadSvg}
\alias{downloadSvg}
\title{Download svg hierarchical dendrogram}
\usage{
downloadSvg(id)
}
\arguments{
\item{id}{namespace}
}
\value{
NA
}
\description{
Download svg hierarchical dendrogram
}
