% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_shinyBS.R
\name{dropNulls}
\alias{dropNulls}
\title{from shinyBS}
\usage{
dropNulls(x)
}
\arguments{
\item{x}{https://github.com/ebailey78/shinyBS/blob/c329f8ce43e44579cafbb16fc3109fb69d403e57/R/updateCollapse.R}
}
\value{
https://github.com/ebailey78/shinyBS/blob/c329f8ce43e44579cafbb16fc3109fb69d403e57/R/updateCollapse.R
}
\description{
from shinyBS
}
