% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_experimentSummary.R
\name{experimentSummary_Server}
\alias{experimentSummary_Server}
\title{experimentSummary_Server}
\usage{
experimentSummary_Server(input, output, session, pool)
}
\arguments{
\item{input}{modules}

\item{output}{modules}

\item{session}{modules}

\item{pool}{modules}
}
\value{
modules
}
\description{
experimentSummary_Server
}
