% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_experimentSummary.R
\name{experimentSummary_UI}
\alias{experimentSummary_UI}
\title{experimentSummary_UI}
\usage{
experimentSummary_UI(id)
}
\arguments{
\item{id}{namespace}
}
\value{
.
}
\description{
experimentSummary_UI
}
