% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_export.R
\name{exportSamples_server}
\alias{exportSamples_server}
\title{exportSamples_server}
\usage{
exportSamples_server(input, output, session, sqlDirectory,
  availableExperiments)
}
\arguments{
\item{input}{shiny}

\item{output}{shiny}

\item{session}{shiny}

\item{sqlDirectory}{sqlDirectory$sqlDirectory}

\item{availableExperiments}{availableExperiments$db}
}
\value{
NA
}
\description{
exportSamples_server
}
