% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_extractBrukerAcqInfo.R
\name{extractBrukerAcquistionInfo}
\alias{extractBrukerAcquistionInfo}
\title{Find a certain Bruker Acqus info}
\usage{
extractBrukerAcquistionInfo(acquisitonInformation, name, type)
}
\arguments{
\item{acquisitonInformation}{parsed acqus}

\item{name}{acqus metadata name}

\item{type}{FUN.VALUE for vapply}
}
\value{
spot position vector
}
\description{
Find a certain Bruker Acqus info
}
