% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readBrukerAcqusFile.r
\name{extractHPCConstants}
\alias{extractHPCConstants}
\title{from https://github.com/sgibb/readBrukerFlexData}
\usage{
extractHPCConstants(hpcStr)
}
\arguments{
\item{hpcStr}{https://github.com/sgibb/readBrukerFlexData}
}
\value{
https://github.com/sgibb/readBrukerFlexData
}
\description{
from https://github.com/sgibb/readBrukerFlexData
}
