% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_mz_files}
\alias{find_mz_files}
\title{Find mzML and mzXML files}
\usage{
find_mz_files(path, recursive = FALSE, full = FALSE)
}
\arguments{
\item{path}{path to search}

\item{recursive}{search directories recursively? T/F}

\item{full}{full.names? T/F}
}
\value{
file paths of found files
}
\description{
Find mzML and mzXML files
}
