% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findIdbacHome.R
\name{findIdbacHome}
\alias{findIdbacHome}
\title{Get default path for IDBac to save experiments to}
\usage{
findIdbacHome()
}
\value{
character path
}
\description{
Get default path for IDBac to save experiments to
}
