% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_findMSconvert.R
\name{findMSconvert}
\alias{findMSconvert}
\title{Find proteowizard's msconvert.exe}
\usage{
findMSconvert(proteoWizardLocation = "")
}
\arguments{
\item{proteoWizardLocation}{optional path to directory containing msconvert.exe}
}
\value{
user's os
}
\description{
Find proteowizard's msconvert.exe
}
