% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findmz}
\alias{findmz}
\title{Find mzML and mzXML files}
\usage{
findmz(inputPath, recursive = FALSE, full = FALSE)
}
\arguments{
\item{inputPath}{path to search}

\item{recursive}{search directories recursively? T/F}

\item{full}{full.names? T/F}
}
\value{
file paths of found files
}
\description{
Find mzML and mzXML files
}
