% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzXML_machineData_readHelpers.r
\name{findRawSHAandFile}
\alias{findRawSHAandFile}
\title{findRawSHAandFile}
\usage{
findRawSHAandFile(singlemzMLpath)
}
\arguments{
\item{singlemzMLpath}{NA}
}
\value{
NA
}
\description{
findRawSHAandFile
}
