% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_update_to_new_version.R
\name{get_db_version}
\alias{get_db_version}
\title{Get the IDBac version that created the database}
\usage{
get_db_version(pool)
}
\arguments{
\item{pool}{pool object}
}
\value{
character version if present, or NULL if error
}
\description{
Get the IDBac version that created the database
}
