% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_getSubTrees.R
\name{get_subtrees}
\alias{get_subtrees}
\title{Get subtree}
\usage{
get_subtrees(
  dend,
  k = NULL,
  h = NULL,
  use_labels_not_values = FALSE,
  order_clusters_as_data = FALSE
)
}
\arguments{
\item{dend}{dendrogram}

\item{k}{numeric scalar (OR a vector) with the number of clusters
the tree should be cut into.}

\item{h}{numeric scalar (OR a vector) with a height where the tree
should be cut.}

\item{use_labels_not_values}{logical, defaults to TRUE. If the actual labels of the
clusters do not matter - and we want to gain speed (say, 10 times faster) -
then use FALSE (gives the "leaves order" instead of their labels.).
This is passed to \code{cutree_1h.dendrogram}.}

\item{order_clusters_as_data}{logical, defaults to TRUE. There are two ways by which
to order the clusters: 1) By the order of the original data. 2) by the order of the
labels in the dendrogram. In order to be consistent with \link[stats]{cutree}, this is set
to TRUE.
This is passed to \code{cutree_1h.dendrogram}.}
}
\value{
dendrogram
}
\description{
Get subtree
}
