% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlManipulators.r
\name{getAllstrain_idsfromSQL}
\alias{getAllstrain_idsfromSQL}
\title{getAllstrain_idsfromSQL}
\usage{
getAllstrain_idsfromSQL(databaseConnection, table)
}
\arguments{
\item{databaseConnection}{NA}

\item{table}{NA}
}
\value{
NA
}
\description{
getAllstrain_idsfromSQL
}
