% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_parseMicrotyper.R
\name{getMicrotyperFiles}
\alias{getMicrotyperFiles}
\title{getMicrotyperFiles}
\usage{
getMicrotyperFiles(dataDirectory = delimitedLocationP())
}
\arguments{
\item{dataDirectory}{User-chosen directory that contains microtyper-exported files directly inside (doesn't search recursively)}
}
\value{
List of files
}
\description{
getMicrotyperFiles
}
