% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseReplicates.R
\name{getPeakData}
\alias{getPeakData}
\title{Retrieve MALDIquant peak objects from an IDBac sqlite database}
\usage{
getPeakData(checkedPool, sampleIDs, protein)
}
\arguments{
\item{checkedPool}{checked out pool}

\item{sampleIDs}{sample IDs of samples to process}

\item{protein}{whether to search SQL for protein or small mol spectra}
}
\value{
unlisted MALDIquant peak objects correspoding to the provided fileshas
}
\description{
Retrieve MALDIquant peak objects from an IDBac sqlite database
}
