% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseProteinReplicates.R
\name{getProteinPeakData}
\alias{getProteinPeakData}
\title{Retrieve MALDIquant peak objects from an IDBac sqlite database}
\usage{
getProteinPeakData(db, fileshas)
}
\arguments{
\item{db}{pool::() connection to the IDBac sqlite dataabse}

\item{fileshas}{the shas for the individual MALDIquant peak objects}
}
\value{
MALDIquant peak objects correspoding to the provided fileshas
}
\description{
Retrieve MALDIquant peak objects from an IDBac sqlite database
}
