% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlManipulators.r
\name{getSmallMolPeakData}
\alias{getSmallMolPeakData}
\title{getSmallMolPeakData}
\usage{
getSmallMolPeakData(db, fileshas)
}
\arguments{
\item{db}{NA}

\item{fileshas}{NA}
}
\value{
NA
}
\description{
getSmallMolPeakData
}
\examples{
NA
}
