% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_connect.R
\name{idbac_connect}
\alias{idbac_connect}
\title{Create a pool connecton, given file names and a path
   If there is no existing database by that name, create one.}
\usage{
idbac_connect(fileName, filePath)
}
\arguments{
\item{fileName}{filename of sqlite database}

\item{filePath}{path of the location of sqlite database, sans filename}
}
\value{
a list of pool connections
}
\description{
Create a pool connecton, given file names and a path
   If there is no existing database by that name, create one.
}
