% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_create_database.R
\name{idbac_create}
\alias{idbac_create}
\title{Title}
\usage{
idbac_create(fileName, filePath)
}
\arguments{
\item{fileName}{file name for new IDBac database}

\item{filePath}{path where database will be created}
}
\value{
path to new database
}
\description{
Title
}
