% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_getDatabaseVersion.R
\name{idbac_db_version}
\alias{idbac_db_version}
\title{Get the IDBac version that created the database}
\usage{
idbac_db_version(pool)
}
\arguments{
\item{pool}{pool object}
}
\value{
numeric_version if present, or NULL if error
}
\description{
Get the IDBac version that created the database
}
