% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dendrogramCreator.R
\name{idbac_dendrogram_creator}
\alias{idbac_dendrogram_creator}
\title{dendrogramCreator}
\usage{
idbac_dendrogram_creator(
  bootstraps = 0L,
  distanceMethod,
  clusteringMethod,
  proteinMatrix
)
}
\arguments{
\item{bootstraps}{integer number of bootstraps}

\item{distanceMethod}{distance method to use}

\item{clusteringMethod}{clustering method to use}

\item{proteinMatrix}{proteinMatrix , rows are samples, cols are peak intensities}
}
\value{
dendrogram
}
\description{
dendrogramCreator
}
