% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_metadataFromId.R
\name{idbac_get_metadata}
\alias{idbac_get_metadata}
\title{Extract a metadata column}
\usage{
idbac_get_metadata(strainID = NULL, metadataColumn = NULL, pool)
}
\arguments{
\item{strainID}{strain_id}

\item{metadataColumn}{metadata column name ('species')}

\item{pool}{pool connection}
}
\value{
data.frame with two columns 'strain_id' and  whatever metadataColumn is requested; rows filtered by ID
}
\description{
Extract a metadata column
}
