% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_idbac_getSpectra.R
\name{idbac_get_spectra}
\alias{idbac_get_spectra}
\title{Get spectra from database}
\usage{
idbac_get_spectra(pool, sample_ids, type, MALDIquant = TRUE)
}
\arguments{
\item{pool}{single pool}

\item{sample_ids}{sample ID to retrieve}

\item{type}{"protein",  "small", or "all"}

\item{MALDIquant}{TRUE/FALSE data.table or MALDIquant spectrum}
}
\value{
single, averaged MALDIquant spectrum
}
\description{
Get spectra from database
}
