% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_dbCreators.R
\name{insertIntoIndividualSpectra}
\alias{insertIntoIndividualSpectra}
\title{Write individual spectra to SQLite}
\usage{
insertIntoIndividualSpectra(
  env,
  XMLinfo,
  pool,
  acquisitionInfo = NULL,
  sampleID
)
}
\arguments{
\item{env}{environment}

\item{XMLinfo}{xmlinfo}

\item{pool}{sqlite pool}

\item{acquisitionInfo}{acquisitionInfo}

\item{sampleID}{sampleID}
}
\value{
writes to sqlite database
}
\description{
Write individual spectra to SQLite
}
