% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbCreators.R
\name{insertIntoMassTable}
\alias{insertIntoMassTable}
\title{Write massTable data to SQLite}
\usage{
insertIntoMassTable(env, userDBCon)
}
\arguments{
\item{env}{environment}

\item{userDBCon}{checked database connection}
}
\value{
nothing, writes to database
}
\description{
Write massTable data to SQLite
}
