% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbCreators.R
\name{insertLocale}
\alias{insertLocale}
\title{Insert current locale info into sql table}
\usage{
insertLocale(userDBCon)
}
\arguments{
\item{userDBCon}{database connection}
}
\value{
side effect
}
\description{
Insert current locale info into sql table
}
