% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_insertMetadataColumns.R
\name{insertMetadataColumns}
\alias{insertMetadataColumns}
\title{Insert columns into IDBac SQLite metadata table}
\usage{
insertMetadataColumns(pool, columnNames)
}
\arguments{
\item{pool}{pool}

\item{columnNames}{character vector of column names to be added}
}
\value{
a single trimmed and binned MALDIquant peak object
}
\description{
Insert columns into IDBac SQLite metadata table
}
