% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelsFromBrushedDendrogram.R
\name{labelsFromBrushedDendrogram}
\alias{labelsFromBrushedDendrogram}
\title{networkViaBrushedDendrogram}
\usage{
labelsFromBrushedDendrogram(dendrogram, brushYmin, brushYmax)
}
\arguments{
\item{dendrogram}{dendrogram}

\item{brushYmin}{double, y-max of user-brush}

\item{brushYmax}{double, y-min of user-brush}
}
\value{
subset of labels that were brushed over in the dendrogram
}
\description{
networkViaBrushedDendrogram
}
