% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_MANS.R
\name{manPageProtDend_Server}
\alias{manPageProtDend_Server}
\title{manPageProtDend_Server}
\usage{
manPageProtDend_Server(
  input,
  output,
  session,
  dendrogram,
  colorByLines,
  cutHeightLines,
  colorByLabels,
  cutHeightLabels,
  plotHeight,
  plotWidth
)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{dendrogram}{dendrogram}

\item{colorByLines}{colorByLines}

\item{cutHeightLines}{cutHeightLines}

\item{colorByLabels}{colorByLabels}

\item{cutHeightLabels}{cutHeightLabels}

\item{plotHeight}{plotHeight}

\item{plotWidth}{plotWidth}
}
\value{
manPageProtDend_Server
}
\description{
manPageProtDend_Server
}
