% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_mans.R
\name{manPageProtDend_UI}
\alias{manPageProtDend_UI}
\title{manPageProtDend_UI}
\usage{
manPageProtDend_UI(id)
}
\arguments{
\item{id}{namespace}
}
\value{
manPageProtDend_UI
}
\description{
manPageProtDend_UI
}
