% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map384Well}
\alias{map384Well}
\title{Create 384-well matrix map}
\usage{
map384Well()
}
\value{
384 well-like matrix, each element in matrix contains its position (eg col 1, row 3 contains "C4")
}
\description{
Create 384-well matrix map
}
