% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mergeDb.R
\name{merge_idbac_experiments}
\alias{merge_idbac_experiments}
\title{Merge all samples from IDBac sqlite databses}
\usage{
merge_idbac_experiments(existing_db_paths, new_db_name, new_db_path)
}
\arguments{
\item{existing_db_paths}{paths to exisitng IDBac experiments/sqlite files}

\item{new_db_name}{new experiment/sqlite database name}

\item{new_db_path}{new experiment/sqlite database path}
}
\description{
Merge all samples from IDBac sqlite databses
}
