% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{metadata_from_id}
\alias{metadata_from_id}
\title{Extract a metadata column}
\usage{
metadata_from_id(strainID, metadataColumn, pool)
}
\arguments{
\item{strainID}{Strain_ID}

\item{metadataColumn}{metadata column name ('Species')}

\item{pool}{pool connection}
}
\value{
data.frame with two columns 'Strain_ID' and  whatever metadataColumn is requested; rows filtered by ID
}
\description{
Extract a metadata column
}
