% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMicrotyper.R
\name{microtyperTomzML}
\alias{microtyperTomzML}
\title{Parse Microtyper files}
\usage{
microtyperTomzML(proteinPaths, proteinNames, smallMolPaths, smallMolNames,
  exportDirectory)
}
\arguments{
\item{proteinPaths}{proteinPaths}

\item{proteinNames}{proteinNames}

\item{smallMolPaths}{smallMolPaths}

\item{smallMolNames}{smallMolNames}

\item{exportDirectory}{NA}
}
\value{
NA
}
\description{
Parse Microtyper files
}
