% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_mirrorPlots.R
\name{mirrorPlotDownload_UI}
\alias{mirrorPlotDownload_UI}
\title{mirrorPlotDownload_UI}
\usage{
mirrorPlotDownload_UI(id)
}
\arguments{
\item{id}{namespace}
}
\value{
NA
}
\description{
mirrorPlotDownload_UI
}
