% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_mirrorPlots.R
\name{mirrorPlotsSettings_UI}
\alias{mirrorPlotsSettings_UI}
\title{mirrorPlotsSettings_UI}
\usage{
mirrorPlotsSettings_UI(id)
}
\arguments{
\item{id}{namespace}
}
\value{
NA
}
\description{
mirrorPlotsSettings_UI
}
