% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirrorPlot_functions.R
\name{mirrorPlotZoom}
\alias{mirrorPlotZoom}
\title{Create zoomed mirror plot}
\usage{
mirrorPlotZoom(mirrorPlotEnv, nameOne, nameTwo, ranges2)
}
\arguments{
\item{mirrorPlotEnv}{data for plot}

\item{nameOne}{sample id of top spectrunm}

\item{nameTwo}{sample id of bottom spectrum}

\item{ranges2}{brushed ranges}
}
\value{

}
\description{
Create zoomed mirror plot
}
