% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_appendToDendrogramLabels.R
\name{modAppendDendLabels_WellPanel}
\alias{modAppendDendLabels_WellPanel}
\title{modAppendDendLabels_WellPanel UI}
\usage{
modAppendDendLabels_WellPanel(ns)
}
\arguments{
\item{ns}{shiny namespace}
}
\value{
shiny ui
}
\description{
modAppendDendLabels_WellPanel UI
}
