% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dendrogram.R
\name{modDendDotsMod_WellPanel}
\alias{modDendDotsMod_WellPanel}
\title{modDendDotsMod_WellPanel UI}
\usage{
modDendDotsMod_WellPanel(ns)
}
\arguments{
\item{ns}{shiny namespace}
}
\value{
shiny ui
}
\description{
modDendDotsMod_WellPanel UI
}
