% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_modifyDendrogramLabels.R
\name{modDendLabels_WellPanel}
\alias{modDendLabels_WellPanel}
\title{modDendLabels_WellPanel UI}
\usage{
modDendLabels_WellPanel(ns)
}
\arguments{
\item{ns}{shiny namespace}
}
\value{
shiny ui
}
\description{
modDendLabels_WellPanel UI
}
