% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_modifyDendrogramLines.R
\name{modDendLines_WellPanel}
\alias{modDendLines_WellPanel}
\title{modDendLines_WellPanel UI}
\usage{
modDendLines_WellPanel(ns)
}
\arguments{
\item{ns}{shiny namespace}
}
\value{
shiny ui
}
\description{
modDendLines_WellPanel UI
}
