% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_peakFuzzifier.R
\name{mquant_bin}
\alias{mquant_bin}
\title{MALDIquant binning}
\usage{
mquant_bin(massStart, massEnd, massList, intensityList, method, tolerance)
}
\arguments{
\item{massStart}{beginning of mass range (as of now, must be smaller than smallest mass)}

\item{massEnd}{end of mass range (as of now, must be smaller than smallest mass)}

\item{massList}{list of mass vectors (eg list(1:10, 1:10))}

\item{intensityList}{list of intesity vectors (eg list(1:10, 1:10))}

\item{method}{
    bin creation rule. \code{"strict"} creates bins never containing two
    or more peaks of the same sample. \code{"relaxed"} allows multiple
    peaks of the same sample in one bin. \code{"reference"} generates bins
    around the mass values from the first \code{\linkS4class{MassPeaks}} object
    in \code{l}.}

\item{tolerance}{
    \code{double}, maximal relative deviation of a peak position (mass) to be
    considered as identical. Must be multiplied by \code{10^-6} for ppm,
    e.g. use \code{tolerance=5e-6} for 5 ppm.
  }
}
\value{
matrix where rows are samples and columns are variables (m/z preojections)
}
\description{
MALDIquant binning
}
\examples{
\dontrun{
masses <- list(Sample_A = c(5000,6000,7000), 
Sample_B = c(5000,6010,7005), Sample_C = c(5000,6010,7005))  
intensities <- list(Sample_A = rep(1, 3), 
Sample_B = rep(1, 3),
 Sample_C = rep(1, 3))
zx <- binnR(massStart = 3000,
            massEnd = 15000,
            ppm = 500,
            massList = masses,
            intensityList = intensities)
            }
}
