% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_mirrorPlots.R
\name{mquantSpecFromSQL}
\alias{mquantSpecFromSQL}
\title{mquantSpecFromSQL}
\usage{
mquantSpecFromSQL(pool, sampleID, protein = FALSE, smallmol = FALSE)
}
\arguments{
\item{pool}{pool}

\item{sampleID}{sampleID}

\item{protein}{TRUE/FALSE to retrieve protein spectra}

\item{smallmol}{TRUE/FALSE to retrieve small molecule spectra}
}
\value{
single, averaged MALDIquant spectrum
}
\description{
mquantSpecFromSQL
}
