% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_peakFuzzifier.r
\name{mQuantToMassVec}
\alias{mQuantToMassVec}
\title{#' Bin MALDI peaks
#'
#' @param peakList list of MALDIquant peak objects
#' @param ppm ppm
#' @param massStart massStart 
#' @param massEnd massEnd 
#'
#' @return list of bin vectors
#' @export
#'
createFuzzyVector <- function(peakList,
                       ppm = 300,
                       massStart = 3000,
                       massEnd = 15000){
  # new binning algo not ready
  #https://github.com/chasemc/IDBacApp/commit/d676e8329e08a149ca11913db663d636b51e6e68#diff-669488b95a9fc327f0603fbfddd4100c
  nams <- names(peakList)
  peakList <- MALDIquant::binPeaks(peakList, method = "relaxed", tolerance = .02)
  peakList <- MALDIquant::intensityMatrix(peakList)
  peakList[is.na(peakList)] <- 0
  rownames(peakList) <- nams
  peakList
}
MALDIquant to Mass Vectors}
\usage{
mQuantToMassVec(peakList)
}
\arguments{
\item{peakList}{MALDIquant peak list}
}
\value{
list of mass vectors
}
\description{
#' Bin MALDI peaks
#'
#' @param peakList list of MALDIquant peak objects
#' @param ppm ppm
#' @param massStart massStart 
#' @param massEnd massEnd 
#'
#' @return list of bin vectors
#' @export
#'
createFuzzyVector <- function(peakList,
                       ppm = 300,
                       massStart = 3000,
                       massEnd = 15000){
  # new binning algo not ready
  #https://github.com/chasemc/IDBacApp/commit/d676e8329e08a149ca11913db663d636b51e6e68#diff-669488b95a9fc327f0603fbfddd4100c
  nams <- names(peakList)
  peakList <- MALDIquant::binPeaks(peakList, method = "relaxed", tolerance = .02)
  peakList <- MALDIquant::intensityMatrix(peakList)
  peakList[is.na(peakList)] <- 0
  rownames(peakList) <- nams
  peakList
}
MALDIquant to Mass Vectors
}
