% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_rawData.R
\name{multipleMaldiPlates}
\alias{multipleMaldiPlates}
\title{multipleMaldiPlates}
\usage{
multipleMaldiPlates(id)
}
\arguments{
\item{id}{NA}
}
\value{
NA
}
\description{
multipleMaldiPlates
}
\examples{
NA
}
