% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mzRpeakSerializer}
\alias{mzRpeakSerializer}
\title{Serialize a list of mzR peak matrices (one column at a time)}
\usage{
mzRpeakSerializer(mzRPeaks, column = c("mass", "intensity"))
}
\arguments{
\item{mzRPeaks}{mzR::Peaks result}

\item{mass}{logical}

\item{intensity}{logical}
}
\value{
list of serials
}
\description{
Serialize a list of mzR peak matrices (one column at a time)
}
