% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_networkFromDF.R
\name{networkFromDF}
\alias{networkFromDF}
\title{Network from data frame}
\usage{
networkFromDF(dataF)
}
\arguments{
\item{dataF}{dataframe}
}
\value{
igraph network
}
\description{
Network from data frame
}
