% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkViaBrushedDendrogram.R
\name{networkViaBrushedDendrogram}
\alias{networkViaBrushedDendrogram}
\title{networkViaBrushedDendrogram}
\usage{
networkViaBrushedDendrogram(dendrogram, brushYmin, brushYmax)
}
\arguments{
\item{dendrogram}{NA}

\item{brushYmin}{NA}

\item{brushYmax}{NA}
}
\value{
NA
}
\description{
networkViaBrushedDendrogram
}
\examples{
NA
}
