% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_updateIdbac.R
\name{new_version_check}
\alias{new_version_check}
\title{Check if there's a new stable version of IDBac}
\usage{
new_version_check()
}
\value{
character string to print to user
}
\description{
Check if there's a new stable version of IDBac
}
