% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_rawData.R
\name{oneMaldiPlateHelpUI}
\alias{oneMaldiPlateHelpUI}
\title{oneMaldiPlateHelpUI}
\usage{
oneMaldiPlateHelpUI(id)
}
\arguments{
\item{id}{NA}
}
\value{
NA
}
\description{
oneMaldiPlateHelpUI
}
\examples{
NA
}
