% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDelimitedMS.R
\name{parseDelimitedMS}
\alias{parseDelimitedMS}
\title{parseDelimitedMS}
\usage{
parseDelimitedMS(proteinPaths, proteinNames, smallMolPaths, smallMolNames,
  exportDirectory, centroid)
}
\arguments{
\item{proteinPaths}{proteinPaths}

\item{proteinNames}{proteinNames}

\item{smallMolPaths}{smallMolPaths}

\item{smallMolNames}{smallMolNames}

\item{exportDirectory}{NA}

\item{centroid}{centroid}
}
\value{
NA
}
\description{
parseDelimitedMS
}
