% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_PCA.R
\name{pca_Server}
\alias{pca_Server}
\title{Plotly PCA Server using irlba}
\usage{
pca_Server(input, output, session, dataframe, namedColors)
}
\arguments{
\item{input}{shiny}

\item{output}{shiny}

\item{session}{shiny}

\item{dataframe}{dataframe that PCA will be performed on}

\item{namedColors}{named vector, names are sample labels, hex colors are the vector}
}
\value{
returns PCA results as reactive data frame
}
\description{
Plotly PCA Server using irlba
}
