% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidimensionalAnalysis.R
\name{pcaCalculation}
\alias{pcaCalculation}
\title{Principle Components Analysis 
Given a data frame whose columns are variables and rows are samples, run Principle Components Analysis (PCA)}
\usage{
pcaCalculation(dataMatrix, logged = FALSE, scaled = TRUE,
  centered = TRUE, missing = 0)
}
\arguments{
\item{dataMatrix}{NA}

\item{logged}{NA}

\item{scaled}{NA}

\item{centered}{NA}

\item{missing}{NA}
}
\value{
a single trimmed and binned MALDIquant peak object
}
\description{
Principle Components Analysis 
Given a data frame whose columns are variables and rows are samples, run Principle Components Analysis (PCA)
}
