% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_PCoA.R
\name{pcoa_Server}
\alias{pcoa_Server}
\title{Plotly pcoa Server using irlba}
\usage{
pcoa_Server(input, output, session, namedColors, distanceMatrix)
}
\arguments{
\item{input}{shiny}

\item{output}{shiny}

\item{session}{shiny}

\item{namedColors}{named vector, names are sample labels, hex colors are the vector}

\item{distanceMatrix}{distance matrix}
}
\value{
returns pcoa results as reactive data frame
}
\description{
Plotly pcoa Server using irlba
}
