% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_multidimensionalAnalyses.R
\name{pcoaCalculation}
\alias{pcoaCalculation}
\title{Principle Coordinates Analysis}
\usage{
pcoaCalculation(distanceMatrix)
}
\arguments{
\item{distanceMatrix}{distance matrix}
}
\value{
3D pcoa data framw
}
\description{
Principle Coordinates Analysis
}
