% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binnR.r
\name{peakBinner}
\alias{peakBinner}
\title{Bin MALDI peaks}
\usage{
peakBinner(peakList, ppm = 300, massStart = 3000, massEnd = 15000)
}
\arguments{
\item{peakList}{list of MALDIquant peak objects}

\item{ppm}{ppm}

\item{massStart}{massStart}

\item{massEnd}{massEnd}
}
\value{
list of bin vectors
}
\description{
Bin MALDI peaks
}
