% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_proteinPeaksToRetainSelections.R
\name{peakRetentionSettings_Server}
\alias{peakRetentionSettings_Server}
\title{peakRetentionSettings_Server}
\usage{
peakRetentionSettings_Server(input, output, session)
}
\arguments{
\item{input}{shiny}

\item{output}{shiny}

\item{session}{shiny}
}
\value{
NA
}
\description{
peakRetentionSettings_Server
}
