% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_proteinPeaksToRetainSelections.R
\name{peakRetentionSettings_UI}
\alias{peakRetentionSettings_UI}
\title{peakRetentionSettings_UI}
\usage{
peakRetentionSettings_UI(id, min_mass = 3000, max_mass = 15000)
}
\arguments{
\item{id}{namespace}

\item{min_mass}{starting min mass}

\item{max_mass}{starting max mass}
}
\value{
NA
}
\description{
peakRetentionSettings_UI
}
