% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_plateMap.R
\name{platemod}
\alias{platemod}
\title{platemod}
\usage{
platemod(input, output, session)
}
\arguments{
\item{input}{shiny module}

\item{output}{shiny module}

\item{session}{shiny module}
}
\value{
rhandsontable
}
\description{
platemod
}
