% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotly_3d_scatter.R
\name{plotly_3d_scatter}
\alias{plotly_3d_scatter}
\title{Create 3d scatter plotly plot}
\usage{
plotly_3d_scatter(data, plotTitle)
}
\arguments{
\item{data}{data frame with data and colors and sample names}

\item{plotTitle}{plot title}
}
\value{
plotly plot
}
\description{
Create 3d scatter plotly plot
}
