% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_popupPlots.R
\name{popupPlot_server}
\alias{popupPlot_server}
\title{popupPlot_server}
\usage{
popupPlot_server(input, output, session, dataFrame)
}
\arguments{
\item{input}{shiny}

\item{output}{shiny}

\item{session}{shiny}

\item{dataFrame}{input dataframe, rows=samples, cols =variables}
}
\value{
NA
}
\description{
popupPlot_server
}
