% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_popupPlots.R
\name{popupPlot_UI}
\alias{popupPlot_UI}
\title{UI module for creating absolute panel popup}
\usage{
popupPlot_UI(id, name)
}
\arguments{
\item{id}{namespace}

\item{name}{name}
}
\value{
PCA UI
}
\description{
UI module for creating absolute panel popup
}
