% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_tsne.R
\name{popupPlotTsne_server}
\alias{popupPlotTsne_server}
\title{popupPlot_server}
\usage{
popupPlotTsne_server(input, output, session, data, namedColors,
  plotTitle = "t-SNE")
}
\arguments{
\item{input}{shiny nput}

\item{output}{shiny output}

\item{session}{shiny session}

\item{data}{data}

\item{namedColors}{named colors to have same colors as dend}

\item{plotTitle}{title of plot}
}
\value{
NA
}
\description{
popupPlot_server
}
