% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_tsne.R
\name{popupPlotTsne_UI}
\alias{popupPlotTsne_UI}
\title{UI module for creating absolute panel popup}
\usage{
popupPlotTsne_UI(id)
}
\arguments{
\item{id}{shiny namespace}
}
\value{
PCA UI
}
\description{
UI module for creating absolute panel popup
}
