% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_processBruker.R
\name{process_bruker}
\alias{process_bruker}
\title{Title}
\usage{
process_bruker(path, pool, sampleMap, ignoreMissing)
}
\arguments{
\item{path}{directory containing Bruker data, shouldn't contain other things
as this searches recursively}

\item{pool}{IDBac database pool... e.g. output of idbac_connect()}

\item{sampleMap}{base::matrix() representing MALDI target plate with sample ids}

\item{ignoreMissing}{TRUE/FALSE whether to ignore target spots missing ids}
}
\value{
NA
}
\description{
Title
}
