% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMZML.R
\name{process_mzml}
\alias{process_mzml}
\title{process_mzml}
\usage{
process_mzml(mzFilePaths, sampleIds, idbacPool, acquisitionInfo)
}
\arguments{
\item{mzFilePaths}{file paths of the mzml files}

\item{sampleIds}{sampleIds that are in the same order as the paths}

\item{idbacPool}{idbac single pool, not a list of pools}

\item{acquisitionInfo}{acquisitionInfo (currently only used when converting from Bruker raw data)}
}
\value{
Nothing direct, creates a sqlite database
}
\description{
process_mzml
}
