% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMZML.R
\name{processMZML}
\alias{processMZML}
\title{processMZML}
\usage{
processMZML(mzFilePaths, sampleIds, sqlDirectory, newExperimentName)
}
\arguments{
\item{mzFilePaths}{file paths of the mzml files}

\item{sampleIds}{sampleIds that are in the same order as the paths}
}
\value{
Nothing direct, creates a sqlite database
}
\description{
processMZML
}
