% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectraToPeaks.R
\name{processProteinSpectra}
\alias{processProteinSpectra}
\title{processProteinSpectra}
\usage{
processProteinSpectra(input)
}
\arguments{
\item{input}{MALDIquant mass spectrum list}
}
\value{
MALDIquant mass spectrum list
}
\description{
processProteinSpectra
}
