% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallMolProcessing.R
\name{processSmallMolSpectra}
\alias{processSmallMolSpectra}
\title{processSmallMolSpectra}
\usage{
processSmallMolSpectra(input)
}
\arguments{
\item{input}{MALDIquant mass spectrum list}
}
\value{
MALDIquant mass spectrum list
}
\description{
processSmallMolSpectra
}
