% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_proteinPeaksToMatrix.r
\name{proteinPeaksToMatrix}
\alias{proteinPeaksToMatrix}
\title{proteinPeaksToMatrix}
\usage{
proteinPeaksToMatrix(bool, proteinPeaks)
}
\arguments{
\item{bool}{NA}

\item{proteinPeaks}{NA}
}
\value{
NA
}
\description{
proteinPeaksToMatrix
}
