% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_proteoWizConvert.R
\name{proteoWizConvert}
\alias{proteoWizConvert}
\title{Convert raw data with msconvert}
\usage{
proteoWizConvert(msconvertPath = "", samplePathList, convertWhere)
}
\arguments{
\item{msconvertPath}{path to MSconvert, if none provided, it will search in Programs folder}

\item{samplePathList}{list where names are sample ids, and list elements are raw file paths}

\item{convertWhere}{directory where temp mzML files are written to}
}
\value{
list(mzFile = convertTo, sampleID = names(convertFrom)))
}
\description{
Convert raw data with msconvert
}
