% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBrukerAcqusFile.r
\name{readAcquFile}
\alias{readAcquFile}
\title{readAcquFile}
\usage{
readAcquFile(fidFile, verbose = FALSE)
}
\arguments{
\item{fidFile}{NA}

\item{verbose}{NA}
}
\value{
NA
}
\description{
readAcquFile
}
\examples{
NA
}
