% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBrukerAcqusFile.r
\name{readAcqusFile}
\alias{readAcqusFile}
\title{Read Bruker Acqus File}
\usage{
readAcqusFile(fidFile, verbose = FALSE)
}
\arguments{
\item{fidFile}{filepath of Bruker Acqus file}

\item{verbose}{should the function be verbose?}
}
\value{
a
}
\description{
Read Bruker Acqus File
}
