% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBrukerTargetSpots.R
\name{readBrukerAcqus}
\alias{readBrukerAcqus}
\title{findBrukerTargetSpots}
\usage{
readBrukerAcqus(brukerDataPath)
}
\arguments{
\item{brukerDataPath}{brukerDataPath}
}
\value{
list of acqus info
}
\description{
findBrukerTargetSpots
}
