% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_retrieveAllStrainIds.r
\name{retrieveAllStrainIds}
\alias{retrieveAllStrainIds}
\title{retrieveAllStrainIds}
\usage{
retrieveAllStrainIds(databaseConnection, table)
}
\arguments{
\item{databaseConnection}{NA}

\item{table}{NA}
}
\value{
NA
}
\description{
retrieveAllStrainIds
}
