% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMicrotyper.R
\name{run_microtyperTomzML}
\alias{run_microtyperTomzML}
\title{run_microtyperTomzML setup and loop}
\usage{
run_microtyperTomzML(proteinPaths = NULL, smallMolPaths = NULL,
  exportDirectory = tempdir())
}
\arguments{
\item{proteinPaths}{proteinPaths}

\item{smallMolPaths}{smallMolPaths}

\item{tempMZDir}{tempMZDir}
}
\value{
NA
}
\description{
run_microtyperTomzML setup and loop
}
