% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierMetaInputModule.R
\name{runDendDots}
\alias{runDendDots}
\title{Run Dend and dots}
\usage{
runDendDots(rawDendrogram, trimdLabsDend, pool, columnID, colors,
  text_shift)
}
\arguments{
\item{rawDendrogram}{NA}

\item{trimdLabsDend}{NA}

\item{pool}{NA}

\item{columnID}{NA}

\item{colors}{NA}

\item{text_shift}{NA}
}
\value{
NA
}
\description{
Run Dend and dots
}
