% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_sampleChooser.R
\name{sampleChooser_server}
\alias{sampleChooser_server}
\title{Server for choosing all samples fromm an IDBac DB}
\usage{
sampleChooser_server(input, output, session, pool,
  whetherProtein = FALSE, allSamples = FALSE, selectedDB)
}
\arguments{
\item{input}{shiny module}

\item{output}{shiny module}

\item{session}{shiny module}

\item{pool}{reactive object that contains a pool to currently-chosen database}

\item{whetherProtein}{T  "allSamples takes precedence}

\item{allSamples}{T/F}
}
\description{
Server for choosing all samples fromm an IDBac DB
}
