% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_sampleChooser.R
\name{sampleChooser_UI}
\alias{sampleChooser_UI}
\title{UI for choosing all samples fromm an IDBac DB}
\usage{
sampleChooser_UI(id)
}
\arguments{
\item{id}{namespace}
}
\value{
UI
}
\description{
UI for choosing all samples fromm an IDBac DB
}
