% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_convertBruker.R
\name{sampleMapSaver}
\alias{sampleMapSaver}
\title{Update sample map reactive value}
\usage{
sampleMapSaver(sampleData, sampleMapReactive)
}
\arguments{
\item{sampleData}{sample map rhandsontable}

\item{sampleMapReactive}{sample map reactive value to save to}
}
\value{
none, side effect
}
\description{
Update sample map reactive value
}
