% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_convertBruker.R
\name{sampleMaptoDF}
\alias{sampleMaptoDF}
\title{Update sample map reactive value}
\usage{
sampleMaptoDF(sampleData)
}
\arguments{
\item{sampleData}{sample map rhandsontable}
}
\value{
none, side effect
}
\description{
Update sample map reactive value
}
