% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_convertBruker.R
\name{sampleMapViewer}
\alias{sampleMapViewer}
\title{Rhandsontable Sample Map}
\usage{
sampleMapViewer(df)
}
\arguments{
\item{df}{reactiveValues data frame input nulledMap384Well()}
}
\value{
rhandsontable
}
\description{
Rhandsontable
}
