% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_addStrainAttributes.r
\name{sampMetaOutput}
\alias{sampMetaOutput}
\title{sampMetaOutput}
\usage{
sampMetaOutput(id, label = "Sample Metadata Input")
}
\arguments{
\item{id}{NA}

\item{label}{NA}
}
\value{
NA
}
\description{
sampMetaOutput
}
