% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathSanitize.R
\name{sanitize}
\alias{sanitize}
\title{Sanitize a filename by removing directory paths and invalid characters}
\usage{
sanitize(filename, replacement = "")
}
\arguments{
\item{filename}{A character vector to be sanitized.}

\item{replacement}{A character vector used to replace invalid characters.}
}
\description{
`path_file_sanitize()` removes the following:
- [Control characters](https://en.wikipedia.org/wiki/C0_and_C1_control_codes)
- [Reserved characters](https://kb.acronis.com/content/39790)
- Unix reserved filenames (`.` and `..`)
- Trailing periods and spaces (invalid on Windows)
- Windows reserved filenames (`CON`, `PRN`, `AUX`, `NUL`, `COM1`, `COM2`,
  `COM3`, COM4, `COM5`, `COM6`, `COM7`, `COM8`, `COM9`, `LPT1`, `LPT2`,
  `LPT3`, `LPT4`, `LPT5`, `LPT6`, LPT7, `LPT8`, and `LPT9`)
The resulting string is then truncated to [255 bytes in length](https://en.wikipedia.org/wiki/Comparison_of_file_systems#Limits)
}
\examples{
# potentially unsafe string
str <- "~/.\u0001ssh/authorized_keys"
sanitize(str)

sanitize("..")
}
\seealso{
<https://www.npmjs.com/package/sanitize-filename>, upon which this
  function is based.
}
