% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_directorySelect.R
\name{selectDirectory_Server}
\alias{selectDirectory_Server}
\title{selectDirectory_Server}
\usage{
selectDirectory_Server(input, output, session, sqlDirectory)
}
\arguments{
\item{input}{shiny}

\item{output}{shiny}

\item{session}{shiny}

\item{sqlDirectory}{sqlDirectory}
}
\value{
NA
}
\description{
selectDirectory_Server
}
