% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_directorySelect.R
\name{selectDirectory_UI}
\alias{selectDirectory_UI}
\title{selectDirectory_UI}
\usage{
selectDirectory_UI(id, label)
}
\arguments{
\item{id}{namespace}

\item{label}{label of the directory action button}
}
\value{
NA
}
\description{
selectDirectory_UI
}
