% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dendInject.R
\name{selectInjections_server}
\alias{selectInjections_server}
\title{Inject samples from another database into dendrogram}
\usage{
selectInjections_server(input, output, session, sqlDirectory,
  availableExperiments, watchMainDb)
}
\arguments{
\item{input}{shiny}

\item{output}{shiny}

\item{session}{shiny}

\item{sqlDirectory}{directory containing sqlite files}

\item{availableExperiments}{available sqlite files}

\item{watchMainDb}{for reacttivity}
}
\value{
NA
}
\description{
Inject samples from another database into dendrogram
}
