% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{serial}
\alias{serial}
\title{serial
   Settings for serializing in IDBac (convert to json)}
\usage{
serial(input)
}
\arguments{
\item{input}{matrix or vector}
}
\value{
JSON
}
\description{
serial
   Settings for serializing in IDBac (convert to json)
}
