% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{serializeXML}
\alias{serializeXML}
\title{Read mzXML, XML and transform to raw character for storing in SQLite}
\usage{
serializeXML(path)
}
\arguments{
\item{path}{xml2 connection}
}
\value{
raw
}
\description{
Read mzXML, XML and transform to raw character for storing in SQLite
}
