% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_smallMolDFtoNetwork.R
\name{small_maldiquant_to_network}
\alias{small_maldiquant_to_network}
\title{Collapse a sample's MALDIquant peak objects into a single peak object}
\usage{
small_maldiquant_to_network(peakList, sampleIDs)
}
\arguments{
\item{peakList}{peakList}

\item{sampleIDs}{sampleIDs}
}
\value{
a single trimmed and binned MALDIquant peak object
}
\description{
Collapse a sample's MALDIquant peak objects into a single peak object
}
