% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_smallMirror.R
\name{smallmirrorPlots_Server}
\alias{smallmirrorPlots_Server}
\title{Title}
\usage{
smallmirrorPlots_Server(input, output, session, workingDB, proteinOrSmall)
}
\arguments{
\item{input}{shiny}

\item{output}{shiny}

\item{session}{shiny}

\item{workingDB}{reactive containing SQLite DB pool}

\item{proteinOrSmall}{values = "proteinPeaks" or "smallMoleculePeaks"}
}
\value{
NA
}
\description{
Title
}
