% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallMorror.R
\name{smallmirrorPlotsSampleSelect_UI}
\alias{smallmirrorPlotsSampleSelect_UI}
\title{mirrorPlotsSettings_UI}
\usage{
smallmirrorPlotsSampleSelect_UI(id)
}
\arguments{
\item{id}{namespace}
}
\value{
NA
}
\description{
mirrorPlotsSettings_UI
}
