% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallMolDFtoNetwork.R
\name{smallMolDFtoNetwork}
\alias{smallMolDFtoNetwork}
\title{Collapse a sample's MALDIquant peak objects into a single peak object}
\usage{
smallMolDFtoNetwork(peakList)
}
\arguments{
\item{peakList}{}
}
\value{
a single trimmed and binned MALDIquant peak object
}
\description{
Collapse a sample's MALDIquant peak objects into a single peak object
}
