% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectraProcessingFunction.r
\name{spectraProcessingFunction}
\alias{spectraProcessingFunction}
\title{Create IDBac SQLite database 
Should work for mzML, mzXML, mgf, and txt data input}
\usage{
spectraProcessingFunction(
  rawDataFilePath,
  sampleID,
  pool,
  acquisitionInfo,
  halfWindowSize
)
}
\arguments{
\item{rawDataFilePath}{filepath of the data}

\item{sampleID}{the sample ID to be read and added to the database}

\item{pool}{database connection (checked out pool)}

\item{acquisitionInfo}{acquisitionInfo (currently only used when converting from Bruker raw data)}

\item{halfWindowSize}{half window size. The resulting
    window reaches from \code{mass[currentIndex-halfWindowSize]} to
    \code{mass[currentIndex+halfWindowSize]} (window size is
    \code{2*halfWindowSize+1}).
    The best size differs depending on the selected \code{method}.}
}
\value{
the peak list modifed by binning then subtractng the matrix sample,
}
\description{
Create IDBac SQLite database 
Should work for mzML, mzXML, mgf, and txt data input
}
