% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlToMaldiquant.R
\name{spectrumMatrixToMALDIqaunt}
\alias{spectrumMatrixToMALDIqaunt}
\title{SQL to MALDIquant}
\usage{
spectrumMatrixToMALDIqaunt(input)
}
\arguments{
\item{input}{matrix/list of martrices , col 1 = masses, col 2 = intensities}
}
\description{
SQL to MALDIquant
}
