% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_sqlTableArchitecture.r
\name{sql_create_locale_table}
\alias{sql_create_locale_table}
\title{SQL code to create the SQLite locale table}
\usage{
sql_create_locale_table(sqlConnection)
}
\arguments{
\item{sqlConnection}{sqlConnection}
}
\value{
SQL code as character
}
\description{
SQL code to create the SQLite locale table
}
