% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlTableArchitecture.r
\name{sql_create_spectra_table}
\alias{sql_create_spectra_table}
\title{SQL code to create the SQLite spectra table}
\usage{
sql_create_spectra_table(sqlConnection)
}
\arguments{
\item{sqlConnection}{sqlConnection}
}
\value{
SQL code as character
}
\description{
SQL code to create the SQLite spectra table
}
