% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlTableArchitecture.r
\name{sql_CreateIndividualSpectra}
\alias{sql_CreateIndividualSpectra}
\title{SQL code to create the SQLite spectra table}
\usage{
sql_CreateIndividualSpectra(sqlConnection)
}
\arguments{
\item{sqlConnection}{sqlConnection}
}
\value{
SQL code as character
}
\description{
SQL code to create the SQLite spectra table
}
