% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlTableArchitecture.r
\name{sql_CreatelLocaleTable}
\alias{sql_CreatelLocaleTable}
\title{SQL code to create the SQLite locale table}
\usage{
sql_CreatelLocaleTable(sqlConnection)
}
\arguments{
\item{sqlConnection}{sqlConnection}
}
\value{
SQL code as character
}
\description{
SQL code to create the SQLite locale table
}
