% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlTableArchitecture.r
\name{sql_CreatemassTable}
\alias{sql_CreatemassTable}
\title{SQL code to create the SQLite mass_index table}
\usage{
sql_CreatemassTable(sqlConnection)
}
\arguments{
\item{sqlConnection}{sqlConnection}
}
\value{
SQL code as character
}
\description{
SQL code to create the SQLite mass_index table
}
