% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlTableArchitecture.r
\name{sql_CreatemetaData}
\alias{sql_CreatemetaData}
\title{SQL code to create the SQLite metaData table}
\usage{
sql_CreatemetaData(sqlConnection)
}
\arguments{
\item{sqlConnection}{sqlConnection}
}
\value{
SQL code as character
}
\description{
SQL code to create the SQLite metaData table
}
