% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlTableArchitecture.r
\name{sql_CreateVersionTable}
\alias{sql_CreateVersionTable}
\title{SQL code to create the SQLite version table}
\usage{
sql_CreateVersionTable(sqlConnection)
}
\arguments{
\item{sqlConnection}{sqlConnection}
}
\value{
SQL code as character
}
\description{
SQL code to create the SQLite version table
}
