% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlTableArchitecture.r
\name{sql_CreatexmlTable}
\alias{sql_CreatexmlTable}
\title{SQL code to create the SQLite xxml table}
\usage{
sql_CreatexmlTable(sqlConnection)
}
\arguments{
\item{sqlConnection}{sqlConnection}
}
\value{
SQL code as character
}
\description{
SQL code to create the SQLite xxml table
}
