% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_dbCreators.R
\name{sql_fill_version_table}
\alias{sql_fill_version_table}
\title{Create version table}
\usage{
sql_fill_version_table(userDBCon)
}
\arguments{
\item{userDBCon}{sqlite connection}
}
\value{
NA
}
\description{
Create version table
}
