% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbCreators.R
\name{sqlCreate_version}
\alias{sqlCreate_version}
\title{Create version table}
\usage{
sqlCreate_version(userDBCon)
}
\arguments{
\item{userDBCon}{sqlite connection}
}
\value{
NA
}
\description{
Create version table
}
