% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_sqlTableArchitecture.r
\name{sqlTableArchitecture}
\alias{sqlTableArchitecture}
\title{sqlTableArchitecture}
\usage{
sqlTableArchitecture(numberScans)
}
\arguments{
\item{numberScans}{NA}
}
\value{
NA
}
\description{
sqlTableArchitecture
}
