% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDataHandlers.R
\name{startingFromMZ}
\alias{startingFromMZ}
\title{When user is starting with mzML/mzXML file(s)}
\usage{
startingFromMZ(chosenDir)
}
\arguments{
\item{chosenDir}{user-chosen directory containing mzML and/or mzXML files}
}
\value{
named character vector: values are mzXX file paths, names are the file names
}
\description{
When user is starting with mzML/mzXML file(s)
}
