% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtractMatrixBlank.R
\name{subtractMatrixBlank}
\alias{subtractMatrixBlank}
\title{Subtract "matrix" sample masses from sample peak lists}
\usage{
subtractMatrixBlank(sampleIds, peakList, binTolerance)
}
\arguments{
\item{sampleIds}{sample IDs corresponding to the peak list}

\item{peakList}{MALDIquant peak objects}

\item{binTolerance}{MALDIquant binPeaks tolerance value (ppm / 10^5)}
}
\value{
the peak list modifed by binning then subtractng the matrix sample, or just the binned peak list if no matrix wsa provided
}
\description{
Subtract "matrix" sample masses from sample peak lists
}
