% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_transferToNewDB.R
\name{transferToNewDB_server}
\alias{transferToNewDB_server}
\title{transferToNewDB_server}
\usage{
transferToNewDB_server(input, output, session, pool, sqlDirectory,
  selectedDB, availableExperiments)
}
\arguments{
\item{input}{.}

\item{output}{.}

\item{session}{.}

\item{pool}{.}

\item{sqlDirectory}{.}

\item{selectedDB}{.}

\item{availableExperiments}{.}
}
\value{
.
}
\description{
transferToNewDB_server
}
