% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_tsne.R
\name{tsne_Server}
\alias{tsne_Server}
\title{Plotly tsne Server using irlba}
\usage{
tsne_Server(input, output, session, namedColors, dataframe)
}
\arguments{
\item{input}{shiny}

\item{output}{shiny}

\item{session}{shiny}

\item{namedColors}{named vector, names are sample labels, hex colors are the vector}

\item{distanceMatrix}{distance matrix}
}
\value{
returns tsne results as reactive data frame
}
\description{
Plotly tsne Server using irlba
}
