% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_proteinClustering.R
\name{ui_proteinClustering}
\alias{ui_proteinClustering}
\title{ui_proteinClustering}
\usage{
ui_proteinClustering(x)
}
\arguments{
\item{x}{NA}
}
\value{
NA
}
\description{
ui_proteinClustering
}
\examples{
NA
}
