% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_update_to_new_version.R
\name{update_database_version}
\alias{update_database_version}
\title{Update IDBac Database from version 1 to 2}
\usage{
update_database_version(pool)
}
\arguments{
\item{pool}{pool object}
}
\value{
NA, side effect
}
\description{
Update IDBac Database from version 1 to 2
}
