% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_shinyBS.R
\name{updateCollapse}
\alias{updateCollapse}
\title{from shinyBS}
\usage{
updateCollapse(session, id, open = NULL, close = NULL, style = NULL)
}
\arguments{
\item{session}{https://github.com/ebailey78/shinyBS/blob/c329f8ce43e44579cafbb16fc3109fb69d403e57/R/updateCollapse.R}

\item{id}{https://github.com/ebailey78/shinyBS/blob/c329f8ce43e44579cafbb16fc3109fb69d403e57/R/updateCollapse.R}

\item{open}{https://github.com/ebailey78/shinyBS/blob/c329f8ce43e44579cafbb16fc3109fb69d403e57/R/updateCollapse.R}

\item{close}{https://github.com/ebailey78/shinyBS/blob/c329f8ce43e44579cafbb16fc3109fb69d403e57/R/updateCollapse.R}

\item{style}{https://github.com/ebailey78/shinyBS/blob/c329f8ce43e44579cafbb16fc3109fb69d403e57/R/updateCollapse.R}
}
\value{
https://github.com/ebailey78/shinyBS/blob/c329f8ce43e44579cafbb16fc3109fb69d403e57/R/updateCollapse.R
}
\description{
from shinyBS
}
