% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_upddateMets.R
\name{updateMeta_server}
\alias{updateMeta_server}
\title{updateMeta_server}
\usage{
updateMeta_server(input, output, session, pool, selectedDB)
}
\arguments{
\item{input}{.}

\item{output}{.}

\item{session}{.}

\item{pool}{.}

\item{selectedDB}{.}
}
\value{
.
}
\description{
updateMeta_server
}
