% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinycssloaders.R
\name{withSpinner}
\alias{withSpinner}
\title{Title}
\usage{
withSpinner(ui_element, type = getOption("spinner.type", default = 1),
  color = getOption("spinner.color", default = "#0275D8"),
  size = getOption("spinner.size", default = 1),
  color.background = getOption("spinner.color.background"),
  custom.css = FALSE, proxy.height = if (grepl("height:\\\\s*\\\\d",
  ui_element)) NULL else "400px")
}
\arguments{
\item{ui_element}{.}

\item{type}{.}

\item{color}{.}

\item{size}{.}

\item{color.background}{.}

\item{custom.css}{.}

\item{proxy.height}{.}
}
\value{
.
}
\description{
Title
}
\examples{
.
}
