% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binnR.r
\name{binnR}
\alias{binnR}
\title{binnR}
\usage{
binnR(vectorList, ppm, massStart, massEnd)
}
\arguments{
\item{vectorList}{NA}

\item{ppm}{NA}

\item{massStart}{NA}

\item{massEnd}{NA}
}
\value{
NA
}
\description{
binnR
}
