% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDataHandlers.R
\name{brukerDataSpotsandPaths}
\alias{brukerDataSpotsandPaths}
\title{Using an excel spreadsheet, get the filepath for msconvert and the user-supplied name}
\usage{
brukerDataSpotsandPaths(brukerDataPath)
}
\arguments{
\item{brukerDataPath}{path to directory containg bruker files}
}
\value{
named list, names are sample IDs, values are paths
}
\description{
Using an excel spreadsheet, get the filepath for msconvert and the user-supplied name
}
