% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dendrogramModifiers.R
\name{changeDendPartColor}
\alias{changeDendPartColor}
\title{changeDendPartColor}
\usage{
changeDendPartColor(
  dendrogram,
  colorBy,
  colorBlindPalette,
  cutHeight = 0,
  chosenK = 1,
  part
)
}
\arguments{
\item{dendrogram}{dendrogram to modify}

\item{colorBy}{color by choosing number of groups, cut height, or metadata}

\item{colorBlindPalette}{colorblind palette}

\item{cutHeight}{height to cut dendrogram}

\item{chosenK}{number of groups for kmeans}

\item{part}{modify dendrogram's labels or lines}
}
\value{
modified dendrogram
}
\description{
changeDendPartColor
}
