% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dendrogramModifiers.R
\name{changeDendPartSize}
\alias{changeDendPartSize}
\title{changeDendPartSize}
\usage{
changeDendPartSize(dendrogram, dendPartSize, part)
}
\arguments{
\item{dendrogram}{dendrogram to modify}

\item{dendPartSize}{numeric passed to change label size or edge widths}

\item{part}{change label size or edge widths}
}
\value{
modified dendrogram
}
\description{
changeDendPartSize
}
