% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_modDendrogramLabels.R
\name{colordendLabels}
\alias{colordendLabels}
\title{colordendLabels}
\usage{
colordendLabels(input, output, session, dendrogram)
}
\arguments{
\item{input}{shiny modules default}

\item{output}{shiny modules default}

\item{session}{shiny modules default}

\item{dendrogram}{dendrogram whose labels are to be modified}
}
\value{
a
}
\description{
colordendLabels
}
