% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compress}
\alias{compress}
\title{compress
  Settings for compressing raw vectors}
\usage{
compress(input, compression = 0)
}
\arguments{
\item{input}{raw vector to compressed}
}
\value{
Raw vector
}
\description{
compress
  Settings for compressing raw vectors
}
