% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_conversion.R
\name{convertDataTab_Server}
\alias{convertDataTab_Server}
\title{convertDataTab_Server}
\usage{
convertDataTab_Server(input, output, session, tempMZDir, sqlDirectory,
  availableExperiments, pwizAvailable)
}
\arguments{
\item{input}{shiny}

\item{output}{shiny}

\item{session}{shiny}

\item{tempMZDir}{directory to create temp mzML if needed}

\item{sqlDirectory}{where to write new SQLite file}

\item{availableExperiments}{update availableExperiments}

\item{pwizAvailable}{whether msconvert was found, logical}
}
\value{
none, updates availableExperiments reactive value though
}
\description{
convertDataTab_Server
}
