% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_convertDelim.R
\name{convertDelim_Server}
\alias{convertDelim_Server}
\title{convertDelim_Server}
\usage{
convertDelim_Server(
  input,
  output,
  session,
  tempMZDir,
  sqlDirectory,
  availableExperiments,
  halfWindowSize
)
}
\arguments{
\item{input}{module}

\item{output}{module}

\item{session}{module}

\item{tempMZDir}{tempMZDir}

\item{sqlDirectory}{sqlDirectory}

\item{availableExperiments}{availableExperiments}

\item{halfWindowSize}{half window size. The resulting
    window reaches from \code{mass[currentIndex-halfWindowSize]} to
    \code{mass[currentIndex+halfWindowSize]} (window size is
    \code{2*halfWindowSize+1}).
    The best size differs depending on the selected \code{method}.}
}
\value{
.
}
\description{
convertDelim_Server
}
