% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_convertBruker.R
\name{convertOneBruker_Server}
\alias{convertOneBruker_Server}
\title{convertOneBruker_Server}
\usage{
convertOneBruker_Server(
  input,
  output,
  session,
  tempMZDir,
  sqlDirectory,
  availableExperiments,
  ...
)
}
\arguments{
\item{input}{NS}

\item{output}{NS}

\item{session}{NS}

\item{tempMZDir}{tempMZDir}

\item{sqlDirectory}{sqlDirectory}

\item{availableExperiments}{availableExperiments}

\item{...}{advanced arguments for MALDIquant, see \code{\link[=processSmallMolSpectra]{processSmallMolSpectra()}} and/or \code{\link[=processProteinSpectra]{processProteinSpectra()}}}
}
\value{
NA
}
\description{
convertOneBruker_Server
}
