% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_convertBruker.R
\name{convertOneBruker_Server}
\alias{convertOneBruker_Server}
\title{convertOneBruker_Server}
\usage{
convertOneBruker_Server(
  input,
  output,
  session,
  tempMZDir,
  sqlDirectory,
  availableExperiments,
  halfWindowSize
)
}
\arguments{
\item{input}{NS}

\item{output}{NS}

\item{session}{NS}

\item{tempMZDir}{tempMZDir}

\item{sqlDirectory}{sqlDirectory}

\item{availableExperiments}{availableExperiments}

\item{halfWindowSize}{half window size. The resulting
    window reaches from \code{mass[currentIndex-halfWindowSize]} to
    \code{mass[currentIndex+halfWindowSize]} (window size is
    \code{2*halfWindowSize+1}).
    The best size differs depending on the selected \code{method}.}
}
\value{
NA
}
\description{
convertOneBruker_Server
}
