% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_copyToNewDatabase.r
\name{copyDB_dbAttach}
\alias{copyDB_dbAttach}
\title{Attach new database to existing database}
\usage{
copyDB_dbAttach(newdbPath, existingDBconn)
}
\arguments{
\item{newdbPath}{newdbPath}

\item{existingDBconn}{existingDBconn}
}
\value{
NA
}
\description{
Attach new database to existing database
}
