% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyToNewDatabase.r
\name{copyToNewDatabase}
\alias{copyToNewDatabase}
\title{Copy from one database to another, selecting by Sample ID}
\usage{
copyToNewDatabase(existingDBPool, newDBPool, newdbName, sampleIDs)
}
\arguments{
\item{existingDBPool}{un-checked pool that samples are being pulled from}

\item{newdbName}{new db name}

\item{sampleIDs}{smaples to transfer}

\item{newdbPath}{path to where new db should be created}
}
\value{
NA
}
\description{
Copy from one database to another, selecting by Sample ID
}
