% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyToNewDatabase.r
\name{copyToNewDatabase}
\alias{copyToNewDatabase}
\title{Copy from one database to another, selecting by Sample ID}
\usage{
copyToNewDatabase(existingDBPool, newDBPool, sampleIDs)
}
\arguments{
\item{existingDBPool}{is a single (not list) pool object referencing the database to transfer from}

\item{newDBPool}{is a single (not list) pool object referencing the database to transfer to}

\item{sampleIDs}{sample IDs to transfer}
}
\value{
Nothing, side effect of creating a new sqlite databse
}
\description{
Copy from one database to another, selecting by Sample ID
}
