% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyToNewDatabase.r
\name{copyToNewDatabase}
\alias{copyToNewDatabase}
\title{Copy from one database to another, selecting by Sample ID}
\usage{
copyToNewDatabase(existingDBPool, newdbPath, sampleIDs)
}
\arguments{
\item{existingDBPool}{NA}

\item{newdbPath}{NA}

\item{sampleIDs}{NA}
}
\value{
NA
}
\description{
Copy from one database to another, selecting by Sample ID
}
