% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbCreators.R
\name{createMetaSQL}
\alias{createMetaSQL}
\title{createMetaSQL}
\usage{
createMetaSQL(sampleID, userDBCon)
}
\arguments{
\item{sampleID}{NA}

\item{userDBCon}{NA}
}
\value{
NA
}
\description{
createMetaSQL
}
