% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPool.R
\name{createNewSQLITEdb}
\alias{createNewSQLITEdb}
\title{Create New, Empty SQLite Database}
\usage{
createNewSQLITEdb(newExperimentName)
}
\arguments{
\item{newExperimentName}{}
}
\description{
Create New, Empty SQLite Database
}
