% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPool.R
\name{createPool}
\alias{createPool}
\title{Create a pool connecton, given file names and a path
If there is no existing database by that name, create one.}
\usage{
createPool(fileName, filePath)
}
\arguments{
\item{fileName}{filename of sqlite database}

\item{filePath}{path of the location of sqlite database, sans filename}
}
\value{
a list of pool connections
}
\description{
Note: Need both name and path due to how users select
}
