% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_dbCreators.R
\name{createSpectraSQL}
\alias{createSpectraSQL}
\title{createSpectraSQL}
\usage{
createSpectraSQL(
  mzML_con,
  pool,
  sampleID,
  XMLinfo,
  smallRangeEnd = 6000,
  acquisitionInfo,
  ...
)
}
\arguments{
\item{mzML_con}{NA}

\item{pool}{sqlite pool}

\item{sampleID}{NA}

\item{XMLinfo}{NA}

\item{smallRangeEnd}{end of mass region for small mol, if m/z above this- will be classified as "protein" spectrum}

\item{acquisitionInfo}{acquisitionInfo (currently only used when converting from Bruker raw data)}

\item{...}{advanced arguments for MALDIquant, see \code{\link[=processSmallMolSpectra]{processSmallMolSpectra()}} and/or \code{\link[=processProteinSpectra]{processProteinSpectra()}}}
}
\value{
writes to sqlite database
}
\description{
createSpectraSQL
}
