% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectraProcessingFunction.r
\name{createSpectraSQL}
\alias{createSpectraSQL}
\title{createSpectraSQL}
\usage{
createSpectraSQL(mzML_con, scanNumber, userDBCon, sampleID, XMLinfo,
  rawDataFilePath)
}
\arguments{
\item{mzML_con}{NA}

\item{scanNumber}{NA}

\item{userDBCon}{NA}

\item{sampleID}{NA}

\item{XMLinfo}{NA}

\item{rawDataFilePath}{NA}
}
\value{
NA
}
\description{
createSpectraSQL
}
