% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_dbCreators.R
\name{createSpectraSQL}
\alias{createSpectraSQL}
\title{createSpectraSQL}
\usage{
createSpectraSQL(
  mzML_con,
  scanNumber,
  userDBCon,
  sampleID,
  XMLinfo,
  smallRangeEnd = 6000,
  acquisitionInfo
)
}
\arguments{
\item{mzML_con}{NA}

\item{scanNumber}{NA}

\item{userDBCon}{NA}

\item{sampleID}{NA}

\item{XMLinfo}{NA}

\item{smallRangeEnd}{end of mass region for small mol, if m/z above this- will be classified as "protein" spectrum}

\item{acquisitionInfo}{acquisitionInfo (currently only used when converting from Bruker raw data)}
}
\value{
NA
}
\description{
createSpectraSQL
}
