% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectraProcessingFunction.r
\name{createXMLSQL}
\alias{createXMLSQL}
\title{createXMLSQL}
\usage{
createXMLSQL(rawDataFilePath, sampleID, userDBCon, mzML_con)
}
\arguments{
\item{rawDataFilePath}{NA}

\item{sampleID}{NA}

\item{userDBCon}{NA}

\item{mzML_con}{NA}
}
\value{
NA
}
\description{
createXMLSQL
}
\examples{
NA
}
