% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_dbCreators.R
\name{createXMLSQL}
\alias{createXMLSQL}
\title{createXMLSQL}
\usage{
createXMLSQL(rawDataFilePath, userDBCon, mzML_con)
}
\arguments{
\item{rawDataFilePath}{NA}

\item{userDBCon}{NA}

\item{mzML_con}{NA}
}
\value{
NA
}
\description{
createXMLSQL
}
