% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dendrogramCreator.R
\name{dendrogramCreator}
\alias{dendrogramCreator}
\title{Title}
\usage{
dendrogramCreator(input, output, session, proteinMatrix)
}
\arguments{
\item{input}{sd}

\item{output}{sd}

\item{session}{sd}

\item{proteinMatrix}{sd}
}
\value{
sd
}
\description{
Title
}
