% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dendrogramCreator.R
\name{dendrogramCreator}
\alias{dendrogramCreator}
\title{dendrogramCreator}
\usage{
dendrogramCreator(input, output, session, proteinMatrix)
}
\arguments{
\item{input}{shiny modules default}

\item{output}{shiny modules default}

\item{session}{shiny modules default}

\item{data}{matrix or dataframe, where columns are variables and rows are samples}
}
\value{
dendrogram
}
\description{
dendrogramCreator
}
