% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_distanceMatrix.r
\name{distMatrix}
\alias{distMatrix}
\title{Create Distance Matrix}
\usage{
distMatrix(data, method)
}
\arguments{
\item{data}{matrix or dataframe, where columns are variables and rows are samples}

\item{method}{distance metric to use}
}
\value{
distance matrix
}
\description{
Create Distance Matrix
}
