% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_updateDatabaseVersion.R
\name{.copy_db}
\alias{.copy_db}
\title{Copy an IDBac database, appending db version number}
\usage{
.copy_db(pool, latest_db_version)
}
\arguments{
\item{pool}{IDBac pool}

\item{latest_db_version}{current database version}
}
\value{
The path for the new database
}
\description{
Copy an IDBac database, appending db version number
}
