% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_idbacRetrievePeaksFromPool.R
\name{.retrieve_peaks_from_pool}
\alias{.retrieve_peaks_from_pool}
\title{Retrieve MALDIquant peak objects from an IDBac sqlite database}
\usage{
.retrieve_peaks_from_pool(pool, sampleIDs, protein)
}
\arguments{
\item{pool}{sqlite pool}

\item{sampleIDs}{sample IDs of samples to process}

\item{protein}{whether to search SQL for protein or small mol spectra}
}
\value{
unlisted MALDIquant peak objects correspoding to the provided fileshas
}
\description{
Retrieve MALDIquant peak objects from an IDBac sqlite database
}
