% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_idbacRetrievePeaksFromPool.R
\name{.retrieve_peaks_from_pool}
\alias{.retrieve_peaks_from_pool}
\title{Retrieve MALDIquant peak objects from an IDBac sqlite database}
\usage{
.retrieve_peaks_from_pool(pool, ids, type, minSNR)
}
\arguments{
\item{pool}{sqlite pool}

\item{ids}{filter retrieved peaks by sample ID (character vector, or NULL to return all)}

\item{type}{"protein",  "small", or "all"}

\item{minSNR}{minimum SNR a a peak must have to be retained}
}
\value{
unlisted MALDIquant peak objects correspoding to the provided fileshas
}
\description{
Retrieve MALDIquant peak objects from an IDBac sqlite database
}
