% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzXML_machineData_readHelpers.r,
%   R/spectraProcessingFunction.r
\name{findAcquisitionInfo}
\alias{findAcquisitionInfo}
\title{findAcquisitionInfo}
\usage{
findAcquisitionInfo(rawFilepaths, manufacturer)

findAcquisitionInfo(rawFilepaths, manufacturer)
}
\arguments{
\item{rawFilepaths}{NA}

\item{manufacturer}{NA}

\item{rawFilepaths}{NA}

\item{manufacturer}{NA}
}
\value{
NA

NA
}
\description{
findAcquisitionInfo

findAcquisitionInfo
}
\examples{
NA
NA
}
