% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBrukerAcquiredSpots.R
\name{findBrukerInfo}
\alias{findBrukerInfo}
\title{Find spots from Bruker Acqus}
\usage{
findBrukerInfo(acquisitonInformation, name)
}
\arguments{
\item{acquisitonInformation}{parsed acqus}

\item{name}{acqus metadata name}
}
\value{
spot position vector
}
\description{
Find spots from Bruker Acqus
}
