% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_findMissingSampleMapIds.R
\name{findMissingSampleMapIds}
\alias{findMissingSampleMapIds}
\title{Find which MALDI-target spots have data but an ID wasn't assigned}
\usage{
findMissingSampleMapIds(spots, sampleMap, ignoreMissing)
}
\arguments{
\item{spots}{spot locations that have data}

\item{sampleMap}{user-input sample names mapped to 384-well plate (16 rows, 24 columns)}

\item{ignoreMissing}{TRUE/FALSE whether to ignore target spots missing ids}
}
\value{
character vector of missing sample IDs or character(0)
}
\description{
Find which MALDI-target spots have data but an ID wasn't assigned
}
