% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msConversion.R
\name{findMSconvert}
\alias{findMSconvert}
\title{findMSconvert}
\usage{
findMSconvert(proteoWizardLocation = character())
}
\arguments{
\item{proteoWizardLocation}{NA}
}
\value{
NA
}
\description{
findMSconvert
}
\examples{
NA
}
