% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMSconvert.R
\name{findMSconvert}
\alias{findMSconvert}
\title{Find msconvert.exe}
\usage{
findMSconvert(proteoWizardLocation = character())
}
\arguments{
\item{proteoWizardLocation}{optional filepath of msconvert.exe to be checked for validity}
}
\value{
filepath of the folder containing msconvert.exe
}
\description{
Find msconvert.exe
}
