% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDataHandlers.R
\name{findMSconvert}
\alias{findMSconvert}
\title{Title}
\usage{
findMSconvert(proteoWizardLocation = "")
}
\arguments{
\item{proteoWizardLocation}{proteoWizardLocation}
}
\value{
user's os
}
\description{
Title
}
