% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlManipulators.r
\name{getAllStrain_IDsfromSQL}
\alias{getAllStrain_IDsfromSQL}
\title{getAllStrain_IDsfromSQL}
\usage{
getAllStrain_IDsfromSQL(databaseConnection, table)
}
\arguments{
\item{databaseConnection}{NA}

\item{table}{NA}
}
\value{
NA
}
\description{
getAllStrain_IDsfromSQL
}
