% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_availableSampleNames.R
\name{idbac_available_samples}
\alias{idbac_available_samples}
\title{Search an IDBac database to see which sample IDs have protein or small molecule data}
\usage{
idbac_available_samples(pool, type)
}
\arguments{
\item{pool}{single IDBac pool connection}

\item{type}{"protein", "small",}
}
\value{
vector of sample names in database with protein or small mol spectra
}
\description{
Search an IDBac database to see which sample IDs have protein or small molecule data
}
