% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_idbac_get_peaks.R
\name{idbac_get_peaks}
\alias{idbac_get_peaks}
\title{Collapse a sample's MALDIquant peak objects into a single peak object}
\usage{
idbac_get_peaks(
  pool,
  sampleIDs,
  minFrequency = 0,
  minNumber = NA,
  lowerMassCutoff,
  upperMassCutoff,
  minSNR,
  tolerance = 0.002,
  type,
  mergeReplicates = TRUE,
  method = "strict",
  verbose = FALSE
)
}
\arguments{
\item{pool}{sqlite pool}

\item{sampleIDs}{sample IDs of samples to process}

\item{minFrequency}{
    \code{double}, remove all peaks which occur in less than
    \code{minFrequency*length(l)} \code{\linkS4class{MassPeaks}}
    objects. It is a relative threshold.
  }

\item{minNumber}{
    \code{double}, remove all peaks which occur in less than
    \code{minNumber} \code{\linkS4class{MassPeaks}}
    objects. It is an absolute threshold.
  }

\item{lowerMassCutoff}{masses below this will be removed from analyses}

\item{upperMassCutoff}{masses above this will be removed from analyses}

\item{minSNR}{minimum SNR a a peak must have to be retained}

\item{tolerance}{
    \code{double}, maximal relative deviation of a peak position (mass) to be
    considered as identical. Must be multiplied by \code{10^-6} for ppm,
    e.g. use \code{tolerance=5e-6} for 5 ppm.
  }

\item{type}{"protein",  "small", or "all"}

\item{mergeReplicates}{should replicates be merged? TRUE/FALSE}

\item{method}{
    bin creation rule. \code{"strict"} creates bins never containing two
    or more peaks of the sampe sample. \code{"relaxed"} allows multiple
    peaks of the same sample in one bin.}

\item{verbose}{should minfreq/minnum warning be displayed}
}
\value{
a single trimmed and binned MALDIquant peak object
}
\description{
Collapse a sample's MALDIquant peak objects into a single peak object
}
