% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_idbac_getPeaks.R
\name{idbac_get_peaks}
\alias{idbac_get_peaks}
\title{Retrieve MALDIquant peak objects from an IDBac sqlite database}
\usage{
idbac_get_peaks(pool, sampleIDs, protein)
}
\arguments{
\item{pool}{sqlite pool}

\item{sampleIDs}{sample IDs of samples to process}

\item{protein}{whether to search SQL for protein or small mol spectra}
}
\value{
unlisted MALDIquant peak objects correspoding to the provided fileshas
}
\description{
Retrieve MALDIquant peak objects from an IDBac sqlite database
}
