% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_idbac_getSpectra.R
\name{idbac_get_spectra}
\alias{idbac_get_spectra}
\title{Get spectra from database}
\usage{
idbac_get_spectra(pool, sampleID, protein = FALSE, smallmol = FALSE)
}
\arguments{
\item{pool}{single pool}

\item{sampleID}{sample ID to retrieve}

\item{protein}{TRUE/FALSE to retrieve protein spectra}

\item{smallmol}{TRUE/FALSE to retrieve small molecule spectra}
}
\value{
single, averaged MALDIquant spectrum
}
\description{
Get spectra from database
}
