% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_updateDatabaseVersion.R
\name{idbac_update_db}
\alias{idbac_update_db}
\title{Update IDBac Database from version 1 to 2}
\usage{
idbac_update_db(pool)
}
\arguments{
\item{pool}{pool object}
}
\value{
NA, side effect
}
\description{
Update IDBac Database from version 1 to 2
}
