% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbCreators.R
\name{insertIntoIndividualSpectra}
\alias{insertIntoIndividualSpectra}
\title{Write individual spectra to SQLite}
\usage{
insertIntoIndividualSpectra(env, XMLinfo, userDBCon, acquisitonInfo)
}
\arguments{
\item{env}{environment}

\item{XMLinfo}{xmlinfo}

\item{userDBCon}{checked database connection}

\item{acquisitonInfo}{acquisitonInfo}
}
\value{
nothing, writes to database
}
\description{
Write individual spectra to SQLite
}
