% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_dbCreators.R
\name{insertIntoMassTable}
\alias{insertIntoMassTable}
\title{Write mass_index data to SQLite}
\usage{
insertIntoMassTable(env, pool)
}
\arguments{
\item{env}{environment}

\item{pool}{sqlite pool}
}
\value{
writes to sqlite database
}
\description{
Write mass_index data to SQLite
}
