% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_MANS.R
\name{MAN_Server}
\alias{MAN_Server}
\title{MAN_Server}
\usage{
MAN_Server(input, output, session, subtractedMatrixBlank, sampleIDs)
}
\arguments{
\item{input}{mod}

\item{output}{mod}

\item{session}{mod}

\item{subtractedMatrixBlank}{subtractedMatrixBlank}

\item{sampleIDs}{sampleIDs}
}
\value{
NA
}
\description{
MAN_Server
}
