% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_MANS.R
\name{MAN_Server}
\alias{MAN_Server}
\title{MAN_Server}
\usage{
MAN_Server(input, output, session, subtractedMatrixBlank, sampleIDs,
  proteinDend)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{subtractedMatrixBlank}{subtractedMatrixBlank}

\item{sampleIDs}{sampleIDs}

\item{proteinDend$dendrogram}{protein dendrogram}
}
\value{
NA
}
\description{
MAN_Server
}
