% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMicrotyper.R
\name{microtyperTomzML}
\alias{microtyperTomzML}
\title{readMicrotyperFiles}
\usage{
microtyperTomzML(key, mzFilePaths)
}
\arguments{
\item{key}{see microtyperTomzML()  list where filepaths are split() by filename}

\item{mzFilePaths}{mzFilePaths mzml file paths}
}
\value{
NA
}
\description{
readMicrotyperFiles
}
