% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_mirrorPlots.R
\name{mirrorPlots_Server}
\alias{mirrorPlots_Server}
\title{Title}
\usage{
mirrorPlots_Server(input, output, session, workingDB, proteinOrSmall)
}
\arguments{
\item{input}{shiny}

\item{output}{shiny}

\item{session}{shiny}

\item{workingDB}{reactive containing SQLite DB pool}

\item{proteinOrSmall}{'>' for protein '<' for small mol}
}
\value{
NA
}
\description{
Title
}
