% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_mirrorPlots.R
\name{mquantSpecFromSQL}
\alias{mquantSpecFromSQL}
\title{mquantSpecFromSQL}
\usage{
mquantSpecFromSQL(checkedPool, sampleID, proteinOrSmall)
}
\arguments{
\item{checkedPool}{checkedPool}

\item{sampleID}{sampleID}

\item{proteinOrSmall}{'>' for protein '<' for small mol}
}
\value{
single, averaged MALDIquant spectrum
}
\description{
mquantSpecFromSQL
}
