% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binnR.r
\name{mQuantToMassVec}
\alias{mQuantToMassVec}
\title{MALDIquant to Mass Vectors}
\usage{
mQuantToMassVec(peakList)
}
\arguments{
\item{peakList}{MALDIquant peak list}
}
\value{
list of mass vectors
}
\description{
MALDIquant to Mass Vectors
}
