% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeSpectrumIntensity.R
\name{normalizeSpectrumIntensity}
\alias{normalizeSpectrumIntensity}
\title{Smooth, remove baseline, and normalize intensity of a spectrum}
\usage{
normalizeSpectrumIntensity(spectrum)
}
\arguments{
\item{spectrum}{spectrum}
}
\value{
normalized MALDIquant massSpectrum
}
\description{
Smooth, remove baseline, and normalize intensity of a spectrum
}
