% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msImportFunctions.R
\name{parseDelimitedMS}
\alias{parseDelimitedMS}
\title{parseDelimitedMS}
\usage{
parseDelimitedMS(proteinDirectory = NULL, smallMolDirectory = NULL,
  exportDirectory)
}
\arguments{
\item{proteinDirectory}{NA}

\item{smallMolDirectory}{NA}

\item{exportDirectory}{NA}
}
\value{
NA
}
\description{
parseDelimitedMS
}
\examples{
NA
}
