% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDelimitedMS.R
\name{parseDelimitedMS}
\alias{parseDelimitedMS}
\title{parseDelimitedMS}
\usage{
parseDelimitedMS(sampleNames = NULL, proteinDirectory = NULL,
  smallMolDirectory = NULL, exportDirectory, centroid)
}
\arguments{
\item{sampleNames}{sample names if provided (not used yet)}

\item{proteinDirectory}{NA}

\item{smallMolDirectory}{NA}

\item{exportDirectory}{NA}
}
\value{
NA
}
\description{
parseDelimitedMS
}
\examples{
NA
}
