% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDelimitedMS.R
\name{parseDelimitedMS}
\alias{parseDelimitedMS}
\title{parseDelimitedMS}
\usage{
parseDelimitedMS(
  proteinPaths,
  proteinNames,
  smallMolPaths,
  smallMolNames,
  exportDirectory,
  centroid
)
}
\arguments{
\item{proteinPaths}{proteinPaths}

\item{proteinNames}{proteinNames}

\item{smallMolPaths}{smallMolPaths}

\item{smallMolNames}{smallMolNames names of the}

\item{exportDirectory}{NA}

\item{centroid}{Do the files contain profile or centroid data (currently implemented as just
one TRUE/FALSE, but could mappply)}
}
\value{
NA
}
\description{
parseDelimitedMS
}
