% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_PCA.R
\name{pca_Server}
\alias{pca_Server}
\title{PCA Server}
\usage{
pca_Server(input, output, session, dataframe, namedColors)
}
\arguments{
\item{input}{shiny}

\item{output}{shiny}

\item{session}{shiny}

\item{dataframe}{dataframe that PCA will be performed on}

\item{namedColors}{named vector, names are sample labels, hex colors are the vector
172-10     172-7     172-1    172-11}
}
\value{
PCA server, returns PCA for report
}
\description{
PCA Server
}
