% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_multidimensionalAnalyses.R
\name{pcaCalculation}
\alias{pcaCalculation}
\title{Principle Components Analysis}
\usage{
pcaCalculation(
  dataMatrix,
  logged = FALSE,
  scaled = FALSE,
  centered = FALSE,
  missing = 0L
)
}
\arguments{
\item{dataMatrix}{data as matrix, columns are variables and rows are samples}

\item{logged}{should the matrix by log10 normalized?}

\item{scaled}{NA}

\item{centered}{NA}

\item{missing}{Missing values should be replaced with...}
}
\value{
a single trimmed and binned MALDIquant peak object
}
\description{
Principle Components Analysis
}
