% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidimensionalAnalysis.R
\name{pcoaCalculation}
\alias{pcoaCalculation}
\title{Principle Coordinates Analysis 
Given a data frame whose columns are variables and rows are samples, run Principle Coordinates Analysis (PCoA)}
\usage{
pcoaCalculation(distanceMatrix)
}
\arguments{
\item{input}{data frame whose columns are variables and rows are samples}
}
\value{
a single trimmed and binned MALDIquant peak object
}
\description{
Principle Coordinates Analysis 
Given a data frame whose columns are variables and rows are samples, run Principle Coordinates Analysis (PCoA)
}
