% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binnR.r
\name{peakBinner}
\alias{peakBinner}
\title{peakBinner}
\usage{
peakBinner(peakList, ppm, massStart, massEnd)
}
\arguments{
\item{peakList}{should be MALDIquant}

\item{ppm}{}

\item{massStart}{}

\item{massEnd}{}
}
\description{
peakBinner
}
