% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binnR.r
\name{peakBinner}
\alias{peakBinner}
\title{peakBinner}
\usage{
peakBinner(peakList, ppm = 300, massStart = NULL, massEnd = NULL)
}
\arguments{
\item{peakList}{should be MALDIquant}

\item{ppm}{ppm}

\item{massStart}{massStart}

\item{massEnd}{massEnd}
}
\value{
list of bin vectors
}
\description{
peakBinner
}
