% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_proteinPeaksToRetainSelections.R
\name{peakRetentionSettings_UI}
\alias{peakRetentionSettings_UI}
\title{peakRetentionSettings_UI}
\usage{
peakRetentionSettings_UI(id, minMass = 3000, maxMass = 15000)
}
\arguments{
\item{id}{namespace}

\item{minMass}{starting min mass}

\item{maxMass}{starting max mass}
}
\value{
NA
}
\description{
peakRetentionSettings_UI
}
