% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_proteinPeaksToRetainSelections.R
\name{peakRetentionSettings_UI}
\alias{peakRetentionSettings_UI}
\title{peakRetentionSettings_UI}
\usage{
peakRetentionSettings_UI(id)
}
\arguments{
\item{id}{namespace}
}
\value{
NA
}
\description{
peakRetentionSettings_UI
}
