% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_plotDendrogram.R
\name{plotDendrogram}
\alias{plotDendrogram}
\title{Function for plotting dendrogram}
\usage{
plotDendrogram(
  dendrogram = dendrogram,
  dendOrPhylo = dendOrPhylo(),
  selectMetaColumn = input$selectMetaColumn,
  appendDendLabels = input$appendDendLabels,
  colorsChosen = colorsChosen(),
  cutHeightLines = input$cutHeightLines,
  colorByLines = input$colorByLines,
  colorByLabels = input$colorByLabels,
  removeDendDots = input$removeDendDots,
  cutHeightLabels = input$cutHeightLabels,
  boots = boots()$bootstraps,
  pool = pool()
)
}
\arguments{
\item{dendrogram}{dendrogram}

\item{dendOrPhylo}{whether to plot leaves as hanging or not}

\item{selectMetaColumn}{plotting dend and dots}

\item{appendDendLabels}{column in metadata sqlite table to append to dend labels}

\item{colorsChosen}{chosen colors for dend and dots}

\item{cutHeightLines}{cutHeightLines}

\item{colorByLines}{colorByLines}

\item{colorByLabels}{colorByLabels}

\item{removeDendDots}{removeDendDots}

\item{cutHeightLabels}{cutHeightLabels}

\item{boots}{plot bootstraps}

\item{pool}{pool reactive}
}
\value{
plot
}
\description{
Function for plotting dendrogram
}
