% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_popupPlots.R
\name{popupPlot_server}
\alias{popupPlot_server}
\title{popupPlot_server}
\usage{
popupPlot_server(input, output, session, dataFrame, namedColors, plotTitle)
}
\arguments{
\item{input}{shiny nput}

\item{output}{shiny output}

\item{session}{shiny session}

\item{dataFrame}{input dataframe, rows = samples, columns = variables}

\item{namedColors}{named colors}

\item{plotTitle}{plot title}
}
\value{
NA
}
\description{
popupPlot_server
}
