% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prioritizer.R
\name{prioritizer}
\alias{prioritizer}
\title{Prioritize Samples}
\usage{
prioritizer(
  pool,
  dendrogram,
  h = NULL,
  k = NULL,
  fractionMetabolitePeaksToRetain,
  minFrequency = 0,
  minNumber = NA,
  smallMolLowerMassCutoff = 200,
  smallMolUpperMassCutoff = 2000,
  smallMolSnr = 10,
  tolerance = 0.002
)
}
\arguments{
\item{pool}{pool connection to IDBac database}

\item{dendrogram}{dendrogram object to be cut, doesn't necessarily have to be protein dendrogram
but labels should be the same as the small molecule data labels}

\item{h}{numeric scalar (OR a vector) with a height where the tree
should be cut.}

\item{k}{numeric scalar (OR a vector) with the number of clusters
the tree should be cut into.}

\item{fractionMetabolitePeaksToRetain}{percent of small molecule peaks to "capture" when selecting isolates}

\item{minFrequency}{percent of peaks that must be present in sample replicates (0 to 1)}

\item{minNumber}{
    \code{double}, remove all peaks which occur in less than
    \code{minNumber} \code{\linkS4class{MassPeaks}}
    objects. It is an absolute threshold.
  }

\item{smallMolLowerMassCutoff}{lower mass cutoff for small molecule spectra}

\item{smallMolUpperMassCutoff}{upper mass cutoff for small molecule spectra}

\item{smallMolSnr}{minimum SNR for small molecule spectra}

\item{tolerance}{
    \code{double}, maximal relative deviation of a peak position (mass) to be
    considered as identical. Must be multiplied by \code{10^-6} for ppm,
    e.g. use \code{tolerance=5e-6} for 5 ppm.
  }
}
\value{
prioritization list
}
\description{
Prioritize Samples
}
