% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prioritizer.R
\name{prioritizer}
\alias{prioritizer}
\title{Prioritize Samples}
\usage{
prioritizer(
  pool,
  dendrogram,
  h = NULL,
  k = NULL,
  percent_metabolite_peaks,
  small_mol_replicates_peakPercentPresence = 0.6,
  small_mol_lowerMassCutoff = 200,
  small_mol_upperMassCutoff = 2000,
  small_mol_SNR = 10,
  tolerance = 0.002
)
}
\arguments{
\item{pool}{pool connection to IDBac database}

\item{h}{numeric scalar (OR a vector) with a height where the tree
should be cut.}

\item{k}{numeric scalar (OR a vector) with the number of clusters
the tree should be cut into.}

\item{percent_metabolite_peaks}{percent of small molecule peaks to "capture" when selecting isolates}

\item{small_mol_replicates_peakPercentPresence}{percent of peaks that must be present in sample replicates (0 to 1)}

\item{small_mol_lowerMassCutoff}{lower mass cutoff for small molecule spectra}

\item{small_mol_upperMassCutoff}{upper mass cutoff for small molecule spectra}

\item{small_mol_SNR}{minimum SNR for small molecule spectra}

\item{tolerance}{
    \code{double}, maximal relative deviation of a peak position (mass) to be
    considered as identical. Must be multiplied by \code{10^-6} for ppm,
    e.g. use \code{tolerance=5e-6} for 5 ppm.
  }
}
\value{
prioritization list
}
\description{
Prioritize Samples
}
