% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMZML.R
\name{processMZML}
\alias{processMZML}
\title{processMZML}
\usage{
processMZML(
  mzFilePaths,
  sampleIds,
  sqlDirectory,
  newExperimentName,
  acquisitionInfo
)
}
\arguments{
\item{mzFilePaths}{file paths of the mzml files}

\item{sampleIds}{sampleIds that are in the same order as the paths}

\item{sqlDirectory}{sqlDirectory}

\item{newExperimentName}{newExperimentName}

\item{acquisitionInfo}{acquisitionInfo (currently only used when converting from Bruker raw data)}
}
\value{
Nothing direct, creates a sqlite database
}
\description{
processMZML
}
