% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectraToPeaks.R
\name{processSmallMolSpectra}
\alias{processSmallMolSpectra}
\title{processSmallMolSpectra}
\usage{
processSmallMolSpectra(input, halfWindowSize = 20)
}
\arguments{
\item{input}{MALDIquant mass spectrum list}

\item{halfWindowSize}{half window size. The resulting
    window reaches from \code{mass[currentIndex-halfWindowSize]} to
    \code{mass[currentIndex+halfWindowSize]} (window size is
    \code{2*halfWindowSize+1}).
    The best size differs depending on the selected \code{method}.}
}
\value{
MALDIquant mass spectrum list
}
\description{
processSmallMolSpectra
}
