% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_processXMLIndSpectra.R
\name{processXMLIndSpectra}
\alias{processXMLIndSpectra}
\title{Process spectra data for input into SQLite}
\usage{
processXMLIndSpectra(spectraImport, smallOrProtein, index, halfWindowSize)
}
\arguments{
\item{spectraImport}{maldiquant spectra list}

\item{smallOrProtein}{"small" or "protein" chooses which processing pipeline is used}

\item{index}{logical vector describing which spectraImport spectra to use}

\item{...}{advanced arguments for MALDIquant, see [IDBacApp::processSmallMolSpectra()], processProteinSpectra()}
}
\value{
environment
}
\description{
Process spectra data for input into SQLite
}
