% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{serial}
\alias{serial}
\title{serial
   Settings for serializing in IDBac}
\usage{
serial(input)
}
\arguments{
\item{input}{object to serialize}
}
\value{
serialized object in binary
}
\description{
serial
   Settings for serializing in IDBac
}
