% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallMolDFtoNetwork.R
\name{smallMolDFtoNetwork}
\alias{smallMolDFtoNetwork}
\title{Collapse a sample's MALDIquant peak objects into a single peak object}
\usage{
smallMolDFtoNetwork(peakList)
}
\arguments{
\item{db}{ool::() connection to the IDBac sqlite dataabse}

\item{fileshas}{the shas for the individual MALDIquant peak objects}

\item{proteinPercentPresence}{value between 0 and 1 (percent presence translated to 0-100%)}

\item{lowerMassCutoff}{masses below this will be removed from analyses}

\item{upperMassCutoff}{masses above this will be removed from analyses}
}
\value{
a single trimmed and binned MALDIquant peak object
}
\description{
Collapse a sample's MALDIquant peak objects into a single peak object
}
