% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_dbCreators.R
\name{sql_fill_locale_table}
\alias{sql_fill_locale_table}
\title{Insert current locale info into sql table}
\usage{
sql_fill_locale_table(pool)
}
\arguments{
\item{pool}{sqlite pool}
}
\value{
writes to sqlite database
}
\description{
Insert current locale info into sql table
}
