% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_dbCreators.R
\name{sql_fill_version_table}
\alias{sql_fill_version_table}
\title{Create version table}
\usage{
sql_fill_version_table(pool)
}
\arguments{
\item{pool}{sqlite pool}
}
\value{
writes to sqlite database
}
\description{
Create version table
}
