% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDataHandlers.R
\name{startingFromBrukerFlex}
\alias{startingFromBrukerFlex}
\title{When user is starting with Bruker Flex file(s)}
\usage{
startingFromBrukerFlex(chosenDir, msconvertPath = "", sampleMap, tempDir)
}
\arguments{
\item{chosenDir}{user-chosen directory containing bruker raw data files}

\item{msconvertPath}{path to MSconvert, if none provided, it will search in Programs folder}

\item{sampleMap}{excel file used for re-naming samples}

\item{tempDir}{directory to temp mzML files are written to}
}
\value{

}
\description{
When user is starting with Bruker Flex file(s)
}
