% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_multidimensionalAnalyses.R
\name{tsneCalculation}
\alias{tsneCalculation}
\title{t-SNE Analysis 
Given a data frame whose columns are variables and rows are samples, run t-SNE}
\usage{
tsneCalculation(dataMatrix, perplexity, theta, iterations)
}
\arguments{
\item{dataMatrix}{NA}

\item{perplexity}{NA}

\item{theta}{NA}

\item{iterations}{NA}
}
\value{
a single trimmed and binned MALDIquant peak object
}
\description{
t-SNE Analysis 
Given a data frame whose columns are variables and rows are samples, run t-SNE
}
